whatwg/20111024.html

<!DOCTYPE html>
<title>IRC logs: freenode / #whatwg / 20111024</title>
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<script src="/irc-logs/s.js"></script>
<h1 id="freenode/whatwg.20111024.log"><a href="/irc-logs/">/irc-logs</a> / freenode / <a href="irc://irc.freenode.net/whatwg">#whatwg</a> / 2011-10-24 / <a href="#end">end</a></h1>
<p>Options: 
<ol id="lines">
 <li id="l-1"><a href="#l-1">#</a> Session Start: Mon Oct 24 00:00:00 2011</li>
 <li id="l-2"><a href="#l-2">#</a> Session Ident: #whatwg</li>
 <li id="l-3" _a="zewt" _r="Philip`"><a href="#l-3">#</a> [00:03] &lt;zewt&gt; Philip`: browsers should definitely expose more internals to their debuggers (bytecode, jit output, etc); browsers seem to treat developers like babies</li>
 <li id="l-4"><a href="#l-4">#</a> [00:04] * Quits: Laster (~quintesse@dhcp-210-195.cs.dartmouth.edu) (Quit: Laster)</li>
 <li id="l-5"><a href="#l-5">#</a> [00:09] * Joins: puppen (~einPaarPu@60-240-245-17.static.tpgi.com.au)</li>
 <li id="l-6"><a href="#l-6">#</a> [00:09] * Quits: AryehGregor (~Simetrica@mediawiki/simetrical) (Ping timeout: 240 seconds)</li>
 <li id="l-7"><a href="#l-7">#</a> [00:10] * Quits: zdobersek (~zan@89.142.251.175) (Quit: Leaving.)</li>
 <li id="l-8"><a href="#l-8">#</a> [00:10] * Joins: nessy (~Adium@74.125.56.18)</li>
 <li id="l-9" _a="Philip`" _r="zewt"><a href="#l-9">#</a> [00:10] &lt;Philip`&gt; zewt: Are any other similar languages any better than JS at this?</li>
 <li id="l-10" _a="zewt"><a href="#l-10">#</a> [00:12] &lt;zewt&gt; you can read back the bytecode of a function in Python (but it doesn't JIT so there's no analogue there, and I'm not sure how documented its bytecode is)</li>
 <li id="l-11" _a="zewt"><a href="#l-11">#</a> [00:13] &lt;zewt&gt; there's also the dis module, but i havn't looked at it much</li>
 <li id="l-12" _a="zewt"><a href="#l-12">#</a> [00:15] &lt;zewt&gt; (i wouldn't suggest that anything about this sort of thing should be anything but browser extensions; not limiting optimizations is far more important)</li>
 <li id="l-13"><a href="#l-13">#</a> [00:20] * Joins: AryehGregor (~Simetrica@cpe-68-175-61-233.nyc.res.rr.com)</li>
 <li id="l-14"><a href="#l-14">#</a> [00:20] * Quits: AryehGregor (~Simetrica@cpe-68-175-61-233.nyc.res.rr.com) (Changing host)</li>
 <li id="l-15"><a href="#l-15">#</a> [00:20] * Joins: AryehGregor (~Simetrica@mediawiki/simetrical)</li>
 <li id="l-16"><a href="#l-16">#</a> [00:22] * Quits: scor (~scor@drupal.org/user/52142/view) (Quit: scor)</li>
 <li id="l-17"><a href="#l-17">#</a> [00:46] * Joins: mpt (~mpt@canonical/mpt)</li>
 <li id="l-18"><a href="#l-18">#</a> [00:48] * Quits: Telling (~unknown@shop3.diku.dk) (Quit: ...)</li>
 <li id="l-19"><a href="#l-19">#</a> [00:55] * Joins: Laster (~quintesse@pat32.dartmouth-secure.border1-cfw.dartmouth.edu)</li>
 <li id="l-20"><a href="#l-20">#</a> [01:00] * Quits: Amorphous (jan@unaffiliated/amorphous) (Ping timeout: 240 seconds)</li>
 <li id="l-21"><a href="#l-21">#</a> [01:00] * Quits: Laster (~quintesse@pat32.dartmouth-secure.border1-cfw.dartmouth.edu) (Quit: Laster)</li>
 <li id="l-22"><a href="#l-22">#</a> [01:01] * Joins: dydx (~dydz@adsl-75-36-190-15.dsl.pltn13.sbcglobal.net)</li>
 <li id="l-23"><a href="#l-23">#</a> [01:15] * Joins: Amorphous (jan@unaffiliated/amorphous)</li>
 <li id="l-24"><a href="#l-24">#</a> [01:21] * Quits: jacobolus (~jacobolus@c-71-198-169-213.hsd1.ca.comcast.net) (Read error: Connection reset by peer)</li>
 <li id="l-25"><a href="#l-25">#</a> [01:28] * Joins: scor (~scor@drupal.org/user/52142/view)</li>
 <li id="l-26"><a href="#l-26">#</a> [01:33] * Quits: mpt (~mpt@canonical/mpt) (Ping timeout: 240 seconds)</li>
 <li id="l-27"><a href="#l-27">#</a> [02:04] * Quits: erlehmann (~erlehmann@dslb-084-062-029-250.pools.arcor-ip.net) (Quit: Ex-Chat)</li>
 <li id="l-28"><a href="#l-28">#</a> [02:08] * Quits: scor (~scor@drupal.org/user/52142/view) (Quit: scor)</li>
 <li id="l-29"><a href="#l-29">#</a> [02:23] * Quits: necolas (~necolas@5e041dd3.bb.sky.com) (Remote host closed the connection)</li>
 <li id="l-30"><a href="#l-30">#</a> [02:32] * Joins: kbrosnan (~kbrosnan@firefox/community/qa/kbrosnan)</li>
 <li id="l-31"><a href="#l-31">#</a> [02:43] * Quits: smaug____ (~chatzilla@MKDCVIII.gprs.sl-laajakaista.fi) (Ping timeout: 255 seconds)</li>
 <li id="l-32"><a href="#l-32">#</a> [02:50] * Joins: mpt (~mpt@71.16.235.2)</li>
 <li id="l-33"><a href="#l-33">#</a> [02:50] * Quits: mpt (~mpt@71.16.235.2) (Changing host)</li>
 <li id="l-34"><a href="#l-34">#</a> [02:50] * Joins: mpt (~mpt@canonical/mpt)</li>
 <li id="l-35"><a href="#l-35">#</a> [02:54] * Joins: yuuki (~kobayashi@58x158x182x50.ap58.ftth.ucom.ne.jp)</li>
 <li id="l-36"><a href="#l-36">#</a> [02:57] * Joins: nattokirai (~nattokira@rtr.mozilla.or.jp)</li>
 <li id="l-37"><a href="#l-37">#</a> [02:58] * Quits: yuuki (~kobayashi@58x158x182x50.ap58.ftth.ucom.ne.jp) (Ping timeout: 258 seconds)</li>
 <li id="l-38"><a href="#l-38">#</a> [02:58] * Joins: yuuki_ (~kobayashi@58x158x182x50.ap58.ftth.ucom.ne.jp)</li>
 <li id="l-39"><a href="#l-39">#</a> [03:07] * Quits: mpt (~mpt@canonical/mpt) (Ping timeout: 240 seconds)</li>
 <li id="l-40"><a href="#l-40">#</a> [03:07] * Joins: mpt (~mpt@canonical/mpt)</li>
 <li id="l-41"><a href="#l-41">#</a> [03:09] * Joins: agektmr (~Adium@2401:fa00:4:1012:fa1e:dfff:fee6:d74e)</li>
 <li id="l-42"><a href="#l-42">#</a> [03:15] * Quits: nessy (~Adium@74.125.56.18) (Quit: Leaving.)</li>
 <li id="l-43"><a href="#l-43">#</a> [03:24] * Quits: connrs (~connrs@conners.plus.com) (Ping timeout: 260 seconds)</li>
 <li id="l-44"><a href="#l-44">#</a> [03:25] * Joins: yutak (~yutak@2401:fa00:4:1008:baac:6fff:fe99:adfb)</li>
 <li id="l-45"><a href="#l-45">#</a> [03:34] * Joins: AlexNRoss (~AleossIRC@unaffiliated/aleoss)</li>
 <li id="l-46"><a href="#l-46">#</a> [03:45] * Joins: jdong_ (~quassel@222.126.155.250)</li>
 <li id="l-47"><a href="#l-47">#</a> [03:50] * Joins: nessy (~Adium@74.125.56.18)</li>
 <li id="l-48"><a href="#l-48">#</a> [03:57] * Quits: _bga (~bga@ppp78-37-199-12.pppoe.avangarddsl.ru) (Read error: Connection reset by peer)</li>
 <li id="l-49"><a href="#l-49">#</a> [03:58] * Quits: mpt (~mpt@canonical/mpt) (Ping timeout: 240 seconds)</li>
 <li id="l-50"><a href="#l-50">#</a> [04:13] * Joins: Evanescence (~Evanescen@122.237.28.26)</li>
 <li id="l-51"><a href="#l-51">#</a> [04:24] * Quits: Evanescence (~Evanescen@122.237.28.26) (Quit: my website: <a href="http://stardiviner.dyndns-blog.com/">http://stardiviner.dyndns-blog.com/</a>)</li>
 <li id="l-52"><a href="#l-52">#</a> [04:24] * Joins: Evanescence (~Evanescen@122.237.28.26)</li>
 <li id="l-53"><a href="#l-53">#</a> [04:26] * Quits: benjoffe_ (~benjoffe_@CPE-121-218-141-115.lnse3.cht.bigpond.net.au) (Remote host closed the connection)</li>
 <li id="l-54"><a href="#l-54">#</a> [04:37] * Joins: jacobolus (~jacobolus@adsl-99-41-54-30.dsl.pltn13.sbcglobal.net)</li>
 <li id="l-55"><a href="#l-55">#</a> [04:39] * Joins: nimbupani (~divyam@219.64.117.145)</li>
 <li id="l-56" _a="nimbupani"><a href="#l-56">#</a> [04:39] * nimbupani is now known as divya</li>
 <li id="l-57"><a href="#l-57">#</a> [05:04] * Joins: rillian_ (~rillian@mist.thaumas.net)</li>
 <li id="l-58"><a href="#l-58">#</a> [05:12] * Joins: nonge_ (~nonge@p508299AE.dip.t-dialin.net)</li>
 <li id="l-59"><a href="#l-59">#</a> [05:16] * Quits: nonge__ (~nonge@p5082A9AD.dip.t-dialin.net) (Ping timeout: 255 seconds)</li>
 <li id="l-60"><a href="#l-60">#</a> [05:38] * Quits: Evanescence (~Evanescen@122.237.28.26) (Quit: my website: <a href="http://stardiviner.dyndns-blog.com/">http://stardiviner.dyndns-blog.com/</a>)</li>
 <li id="l-61"><a href="#l-61">#</a> [05:40] * Joins: Evanescence (~Evanescen@122.237.28.26)</li>
 <li id="l-62"><a href="#l-62">#</a> [05:47] * Quits: yutak (~yutak@2401:fa00:4:1008:baac:6fff:fe99:adfb) (Ping timeout: 240 seconds)</li>
 <li id="l-63"><a href="#l-63">#</a> [06:10] * Quits: AlexNRoss (~AleossIRC@unaffiliated/aleoss) (Ping timeout: 260 seconds)</li>
 <li id="l-64"><a href="#l-64">#</a> [06:16] * Quits: payman (~payman@pat.se.opera.com) (Read error: Operation timed out)</li>
 <li id="l-65"><a href="#l-65">#</a> [06:16] * Quits: zewt (~x@c-24-62-196-44.hsd1.ma.comcast.net) (Ping timeout: 252 seconds)</li>
 <li id="l-66"><a href="#l-66">#</a> [06:16] * Quits: krijn (u2319@gateway/web/irccloud.com/x-gokkmbxvqpqshyjb) (Read error: Operation timed out)</li>
 <li id="l-67"><a href="#l-67">#</a> [06:16] * Quits: timeless (u4015@firefox/developer/timeless) (Read error: Operation timed out)</li>
 <li id="l-68"><a href="#l-68">#</a> [06:16] * Quits: jmb (~jmb@mail.parsifal.org.uk) (Write error: Broken pipe)</li>
 <li id="l-69"><a href="#l-69">#</a> [06:16] * Quits: lensco (u4054@gateway/web/irccloud.com/x-mncsnjxxitamorlr) (Read error: Operation timed out)</li>
 <li id="l-70"><a href="#l-70">#</a> [06:16] * Quits: bobylito (u3929@gateway/web/irccloud.com/x-bdavhajeyqemilnv) (Write error: Broken pipe)</li>
 <li id="l-71"><a href="#l-71">#</a> [06:16] * Joins: zewt (~x@c-24-62-196-44.hsd1.ma.comcast.net)</li>
 <li id="l-72"><a href="#l-72">#</a> [06:16] * Joins: jmb (~jmb@mail.parsifal.org.uk)</li>
 <li id="l-73"><a href="#l-73">#</a> [06:17] * Quits: kinetik (~kinetik@121.98.132.55) (Ping timeout: 252 seconds)</li>
 <li id="l-74"><a href="#l-74">#</a> [06:18] * Joins: kinetik (~kinetik@121.98.132.55)</li>
 <li id="l-75"><a href="#l-75">#</a> [06:18] * Joins: zcorpan (~zcorpan@c-699de355.410-6-64736c14.cust.bredbandsbolaget.se)</li>
 <li id="l-76"><a href="#l-76">#</a> [06:18] * Joins: payman (~payman@pat.se.opera.com)</li>
 <li id="l-77"><a href="#l-77">#</a> [06:19] * Joins: timeless (u4015@gateway/web/irccloud.com/x-yjcvbmfjxrhilazl)</li>
 <li id="l-78"><a href="#l-78">#</a> [06:19] * Quits: timeless (u4015@gateway/web/irccloud.com/x-yjcvbmfjxrhilazl) (Changing host)</li>
 <li id="l-79"><a href="#l-79">#</a> [06:19] * Joins: timeless (u4015@firefox/developer/timeless)</li>
 <li id="l-80"><a href="#l-80">#</a> [06:23] * Joins: bobylito (u3929@gateway/web/irccloud.com/x-zduvlymsahabqlar)</li>
 <li id="l-81"><a href="#l-81">#</a> [06:23] * Quits: zcorpan (~zcorpan@c-699de355.410-6-64736c14.cust.bredbandsbolaget.se) (Quit: zcorpan)</li>
 <li id="l-82"><a href="#l-82">#</a> [06:24] * Joins: lensco (u4054@gateway/web/irccloud.com/x-vsamtqoxgwexuqqw)</li>
 <li id="l-83"><a href="#l-83">#</a> [06:26] * Quits: Evanescence (~Evanescen@122.237.28.26) (Quit: my website: <a href="http://stardiviner.dyndns-blog.com/">http://stardiviner.dyndns-blog.com/</a>)</li>
 <li id="l-84"><a href="#l-84">#</a> [06:27] * Joins: Evanescence (~Evanescen@122.237.28.26)</li>
 <li id="l-85"><a href="#l-85">#</a> [06:31] * Joins: krijn (u2319@gateway/web/irccloud.com/x-iqzzzakvphwdppeu)</li>
 <li id="l-86"><a href="#l-86">#</a> [06:31] * Quits: divya (~divyam@219.64.117.145) (Ping timeout: 258 seconds)</li>
 <li id="l-87"><a href="#l-87">#</a> [06:32] * Joins: nimbupani (~divyam@219.64.117.145)</li>
 <li id="l-88" _a="nimbupani"><a href="#l-88">#</a> [06:32] * nimbupani is now known as divya</li>
 <li id="l-89"><a href="#l-89">#</a> [06:46] * Quits: kennyluck (~kennyluck@114-43-119-208.dynamic.hinet.net) (Quit: kennyluck)</li>
 <li id="l-90"><a href="#l-90">#</a> [06:48] * Quits: divya (~divyam@219.64.117.145) (Ping timeout: 260 seconds)</li>
 <li id="l-91"><a href="#l-91">#</a> [06:53] * Joins: nimbupani (~divyam@219.64.117.145)</li>
 <li id="l-92" _a="nimbupani"><a href="#l-92">#</a> [06:53] * nimbupani is now known as divya</li>
 <li id="l-93"><a href="#l-93">#</a> [06:59] * Quits: rillian_ (~rillian@mist.thaumas.net) (Remote host closed the connection)</li>
 <li id="l-94"><a href="#l-94">#</a> [06:59] * Joins: yutak (~yutak@2401:fa00:4:1000:baac:6fff:fe99:adfb)</li>
 <li id="l-95"><a href="#l-95">#</a> [07:08] * Quits: gavin_ (~gavin@76.14.70.183) (Quit: Lost terminal)</li>
 <li id="l-96"><a href="#l-96">#</a> [07:08] * Joins: connrs (~connrs@conners.plus.com)</li>
 <li id="l-97" _a="heycam|away"><a href="#l-97">#</a> [07:11] * heycam|away is now known as heycam</li>
 <li id="l-98"><a href="#l-98">#</a> [07:27] * Joins: Areks (~Areks@rs.gridnine.com)</li>
 <li id="l-99"><a href="#l-99">#</a> [07:28] * Quits: ezoe (~ezoe@203-140-89-204f1.kyt1.eonet.ne.jp) (Ping timeout: 240 seconds)</li>
 <li id="l-100"><a href="#l-100">#</a> [07:41] * Quits: Evanescence (~Evanescen@122.237.28.26) (Quit: my website: <a href="http://stardiviner.dyndns-blog.com/">http://stardiviner.dyndns-blog.com/</a>)</li>
 <li id="l-101"><a href="#l-101">#</a> [07:43] * Joins: Evanescence (~Evanescen@122.237.28.26)</li>
 <li id="l-102"><a href="#l-102">#</a> [07:44] * Quits: divya (~divyam@219.64.117.145) (Ping timeout: 255 seconds)</li>
 <li id="l-103"><a href="#l-103">#</a> [07:49] * Joins: Rich_Clark (~chatzilla@94-193-82-82.zone7.bethere.co.uk)</li>
 <li id="l-104"><a href="#l-104">#</a> [07:58] * Quits: yutak (~yutak@2401:fa00:4:1000:baac:6fff:fe99:adfb) (Quit: Ex-Chat)</li>
 <li id="l-105" _a="heycam"><a href="#l-105">#</a> [07:58] * heycam is now known as heycam|away</li>
 <li id="l-106"><a href="#l-106">#</a> [07:59] * Quits: dydx (~dydz@adsl-75-36-190-15.dsl.pltn13.sbcglobal.net) (Quit: dydx)</li>
 <li id="l-107"><a href="#l-107">#</a> [08:01] * Joins: nimbupani (~divyam@219.64.117.145)</li>
 <li id="l-108" _a="nimbupani"><a href="#l-108">#</a> [08:01] * nimbupani is now known as divya</li>
 <li id="l-109"><a href="#l-109">#</a> [08:04] * Joins: yutak (~yutak@2401:fa00:4:1000:baac:6fff:fe99:adfb)</li>
 <li id="l-110"><a href="#l-110">#</a> [08:04] * Joins: ezoe (~ezoe@203-140-88-62f1.kyt1.eonet.ne.jp)</li>
 <li id="l-111"><a href="#l-111">#</a> [08:07] * Quits: Areks (~Areks@rs.gridnine.com) (Quit: KVIrc 4.0.2 Insomnia <a href="http://www.kvirc.net/">http://www.kvirc.net/</a>)</li>
 <li id="l-112"><a href="#l-112">#</a> [08:09] * Quits: nessy (~Adium@74.125.56.18) (Quit: Leaving.)</li>
 <li id="l-113"><a href="#l-113">#</a> [08:19] * Quits: divya (~divyam@219.64.117.145) (Ping timeout: 245 seconds)</li>
 <li id="l-114"><a href="#l-114">#</a> [08:19] * Joins: nimbupani (~divyam@219.64.117.145)</li>
 <li id="l-115" _a="nimbupani"><a href="#l-115">#</a> [08:19] * nimbupani is now known as divya</li>
 <li id="l-116"><a href="#l-116">#</a> [08:34] * Joins: zcorpan (~zcorpan@c-699de355.410-6-64736c14.cust.bredbandsbolaget.se)</li>
 <li id="l-117"><a href="#l-117">#</a> [08:35] * Joins: tomasf (~tomasf@77.72.97.5.c.fiberdirekt.net)</li>
 <li id="l-118"><a href="#l-118">#</a> [08:58] * Joins: virtuelv (~virtuelv_@pat-tdc.opera.com)</li>
 <li id="l-119"><a href="#l-119">#</a> [09:10] * Joins: zdobersek (~zan@89.142.251.175)</li>
 <li id="l-120"><a href="#l-120">#</a> [09:15] * Quits: manu-db (~msporny@digitalbazaar.com) (Read error: Operation timed out)</li>
 <li id="l-121"><a href="#l-121">#</a> [09:16] * Joins: manu-db (~msporny@digitalbazaar.com)</li>
 <li id="l-122"><a href="#l-122">#</a> [09:20] * Joins: brucel (~brucel@cpc4-smal11-2-0-cust879.perr.cable.virginmedia.com)</li>
 <li id="l-123"><a href="#l-123">#</a> [09:23] * Quits: ezoe (~ezoe@203-140-88-62f1.kyt1.eonet.ne.jp) (Ping timeout: 252 seconds)</li>
 <li id="l-124"><a href="#l-124">#</a> [09:24] * Joins: FlorianX (~Florian_S@p4FE2DDEE.dip.t-dialin.net)</li>
 <li id="l-125"><a href="#l-125">#</a> [09:39] * Quits: divya (~divyam@219.64.117.145) (Ping timeout: 240 seconds)</li>
 <li id="l-126"><a href="#l-126">#</a> [09:48] * Quits: jacobolus (~jacobolus@adsl-99-41-54-30.dsl.pltn13.sbcglobal.net) (Remote host closed the connection)</li>
 <li id="l-127"><a href="#l-127">#</a> [09:55] * Joins: zdobersek1 (~zan@89.142.251.175)</li>
 <li id="l-128"><a href="#l-128">#</a> [09:55] * Quits: zdobersek (~zan@89.142.251.175) (Read error: Connection reset by peer)</li>
 <li id="l-129"><a href="#l-129">#</a> [10:11] * Quits: nattokirai (~nattokira@rtr.mozilla.or.jp) (Quit: nattokirai)</li>
 <li id="l-130"><a href="#l-130">#</a> [10:28] * Joins: FlorianX1 (~Florian_S@p4FE2CCFB.dip.t-dialin.net)</li>
 <li id="l-131"><a href="#l-131">#</a> [10:30] * Joins: jacobolus (~jacobolus@c-71-198-169-213.hsd1.ca.comcast.net)</li>
 <li id="l-132"><a href="#l-132">#</a> [10:31] * Joins: akamike (~akamike@94-193-106-14.zone7.bethere.co.uk)</li>
 <li id="l-133"><a href="#l-133">#</a> [10:32] * Quits: FlorianX (~Florian_S@p4FE2DDEE.dip.t-dialin.net) (Ping timeout: 260 seconds)</li>
 <li id="l-134"><a href="#l-134">#</a> [10:32] * Quits: shepazu (~shepazu@108-70-132-46.lightspeed.rlghnc.sbcglobal.net) (Quit: shepazu)</li>
 <li id="l-135"><a href="#l-135">#</a> [10:49] * Joins: Areks (~Areks@rs.gridnine.com)</li>
 <li id="l-136"><a href="#l-136">#</a> [10:56] * Joins: zdobersek (~zan@89.142.251.175)</li>
 <li id="l-137"><a href="#l-137">#</a> [10:56] * Quits: zdobersek1 (~zan@89.142.251.175) (Read error: Connection reset by peer)</li>
 <li id="l-138"><a href="#l-138">#</a> [10:56] * Joins: rtuin (~rtuin@213.125.175.250)</li>
 <li id="l-139"><a href="#l-139">#</a> [11:09] * Joins: mishunov (~spliter@77.88.72.162)</li>
 <li id="l-140"><a href="#l-140">#</a> [11:09] * Joins: annevk (~annevk@EM114-48-14-90.pool.e-mobile.ne.jp)</li>
 <li id="l-141"><a href="#l-141">#</a> [11:37] * Quits: annevk (~annevk@EM114-48-14-90.pool.e-mobile.ne.jp) (Quit: annevk)</li>
 <li id="l-142"><a href="#l-142">#</a> [11:54] * Quits: zdobersek (~zan@89.142.251.175) (Ping timeout: 255 seconds)</li>
 <li id="l-143"><a href="#l-143">#</a> [11:54] * Joins: smaug____ (~chatzilla@dsl-hkibrasgw3-ff5ac100-203.dhcp.inet.fi)</li>
 <li id="l-144"><a href="#l-144">#</a> [11:55] * Joins: annevk (~annevk@EM114-48-14-90.pool.e-mobile.ne.jp)</li>
 <li id="l-145"><a href="#l-145">#</a> [11:58] * Joins: smaug_____ (~chatzilla@GYGMMCLXXXVI.gprs.sl-laajakaista.fi)</li>
 <li id="l-146"><a href="#l-146">#</a> [11:59] * Quits: smaug____ (~chatzilla@dsl-hkibrasgw3-ff5ac100-203.dhcp.inet.fi) (Ping timeout: 240 seconds)</li>
 <li id="l-147" _a="smaug_____"><a href="#l-147">#</a> [11:59] * smaug_____ is now known as smaug____</li>
 <li id="l-148" _a="annevk" _r="zcorpan"><a href="#l-148">#</a> [12:02] &lt;annevk&gt; zcorpan, hey did you think about a test format?</li>
 <li id="l-149" _a="annevk"><a href="#l-149">#</a> [12:02] &lt;annevk&gt; it would be great to have something like the HTML stuff</li>
 <li id="l-150" _a="annevk" _r="input"><a href="#l-150">#</a> [12:02] &lt;annevk&gt; input, output, and input contains errors</li>
 <li id="l-151" _a="annevk"><a href="#l-151">#</a> [12:03] &lt;annevk&gt; and I guess we should have some shortcuts for input, so that you e.g. do not need to provide the signature, or cue time information</li>
 <li id="l-152" _a="annevk"><a href="#l-152">#</a> [12:03] &lt;annevk&gt; (in such a scenario they would just be pre-filled)</li>
 <li id="l-153"><a href="#l-153">#</a> [12:06] * Joins: karlcow (~karl@nerval.la-grange.net)</li>
 <li id="l-154"><a href="#l-154">#</a> [12:07] * Joins: MikeSmith (~MikeSmith@EM114-48-14-90.pool.e-mobile.ne.jp)</li>
 <li id="l-155"><a href="#l-155">#</a> [12:10] * Joins: zdobersek (~zan@89.142.251.175)</li>
 <li id="l-156" _a="annevk"><a href="#l-156">#</a> [12:11] &lt;annevk&gt; lol</li>
 <li id="l-157" _a="annevk"><a href="#l-157">#</a> [12:11] &lt;annevk&gt; webrtc fork uses respec</li>
 <li id="l-158" class="flagged" _a="smaug____" _r="so"><a href="#l-158">#</a> [12:11] &lt;smaug____&gt; so, of the pages which nightly telemetry users have loaded, 2% has had mutation listeners :/</li>
 <li id="l-159" _a="annevk"><a href="#l-159">#</a> [12:12] &lt;annevk&gt; with the silly &lt;dl&gt; way of defining methods</li>
 <li id="l-160" _a="smaug____"><a href="#l-160">#</a> [12:12] &lt;smaug____&gt; of course it is possible that gmail or some such has mutation listeners</li>
 <li id="l-161" _a="smaug____"><a href="#l-161">#</a> [12:12] &lt;smaug____&gt; and every new message loads a new page or something...</li>
 <li id="l-162" _a="annevk"><a href="#l-162">#</a> [12:13] &lt;annevk&gt; is a different fragment identifier a new page?</li>
 <li id="l-163" _a="smaug____"><a href="#l-163">#</a> [12:14] &lt;smaug____&gt; in this case? this is not based on urls, but effectively js windows</li>
 <li id="l-164" _a="smaug____"><a href="#l-164">#</a> [12:14] &lt;smaug____&gt; s/js/dom/</li>
 <li id="l-165" _a="smaug____"><a href="#l-165">#</a> [12:15] &lt;smaug____&gt; different pages get a new window</li>
 <li id="l-166" _a="smaug____"><a href="#l-166">#</a> [12:15] &lt;smaug____&gt; (expect in certain cases with about:blank)</li>
 <li id="l-167" _a="smaug____"><a href="#l-167">#</a> [12:15] &lt;smaug____&gt; except</li>
 <li id="l-168" _a="annevk" _r="ah"><a href="#l-168">#</a> [12:16] &lt;annevk&gt; ah, I think Gmail is pretty much all one window</li>
 <li id="l-169" _a="annevk" _r="oh"><a href="#l-169">#</a> [12:16] &lt;annevk&gt; oh, or maybe not</li>
 <li id="l-170" _a="smaug____"><a href="#l-170">#</a> [12:16] &lt;smaug____&gt; it is using iframes everywhere</li>
 <li id="l-171" _a="annevk" _r="yeah"><a href="#l-171">#</a> [12:16] &lt;annevk&gt; yeah, forgot about that, ignore me</li>
 <li id="l-172" _a="smaug____"><a href="#l-172">#</a> [12:17] &lt;smaug____&gt; I should probably count usage per domain</li>
 <li id="l-173"><a href="#l-173">#</a> [12:17] * Quits: MikeSmith (~MikeSmith@EM114-48-14-90.pool.e-mobile.ne.jp) (Quit: MikeSmith)</li>
 <li id="l-174" _a="annevk"><a href="#l-174">#</a> [12:17] &lt;annevk&gt; gotta catch a train</li>
 <li id="l-175" _a="annevk"><a href="#l-175">#</a> [12:17] &lt;annevk&gt; yeah</li>
 <li id="l-176"><a href="#l-176">#</a> [12:17] * Quits: annevk (~annevk@EM114-48-14-90.pool.e-mobile.ne.jp) (Quit: annevk)</li>
 <li id="l-177"><a href="#l-177">#</a> [12:19] * Joins: bga_ (~bga@ppp78-37-199-12.pppoe.avangarddsl.ru)</li>
 <li id="l-178" _a="foolip" _r="zcorpan"><a href="#l-178">#</a> [12:20] &lt;foolip&gt; zcorpan, do you have any suggestions for what to do about the &quot;separating cues&quot; issue?</li>
 <li id="l-179"><a href="#l-179">#</a> [12:22] * Joins: annevk (~annevk@EM114-48-14-90.pool.e-mobile.ne.jp)</li>
 <li id="l-180" _a="annevk"><a href="#l-180">#</a> [12:24] &lt;annevk&gt; I wonder how wildly different Opera's &lt;track&gt; implementation is compared to mine</li>
 <li id="l-181"><a href="#l-181">#</a> [12:27] * Joins: MikeSmith (~MikeSmith@EM114-48-14-90.pool.e-mobile.ne.jp)</li>
 <li id="l-182"><a href="#l-182">#</a> [12:37] * Joins: erlehmann (~erlehmann@dslb-084-062-029-250.pools.arcor-ip.net)</li>
 <li id="l-183" _a="karlcow"><a href="#l-183">#</a> [12:38] &lt;karlcow&gt; <a href="http://research.microsoft.com/apps/pubs/default.aspx?id=155295">http://research.microsoft.com/apps/pubs/default.aspx?id=155295</a></li>
 <li id="l-184" _a="karlcow"><a href="#l-184">#</a> [12:38] &lt;karlcow&gt; Refactoring the Web Interface</li>
 <li id="l-185" _a="karlcow"><a href="#l-185">#</a> [12:38] &lt;karlcow&gt; John R. Douceur, Jon Howell, Bryan Parno, and Michael Walfish</li>
 <li id="l-186" _a="karlcow"><a href="#l-186">#</a> [12:38] &lt;karlcow&gt; 24 October 2011</li>
 <li id="l-187" _a="karlcow"><a href="#l-187">#</a> [12:38] &lt;karlcow&gt; <a href="http://research.microsoft.com/pubs/155295/SOSP2011Poster.pdf">http://research.microsoft.com/pubs/155295/SOSP2011Poster.pdf</a> and <a href="http://research.microsoft.com/pubs/155295/SOSP2011PosterAbstract.pdf">http://research.microsoft.com/pubs/155295/SOSP2011PosterAbstract.pdf</a></li>
 <li id="l-188"><a href="#l-188">#</a> [12:42] * Quits: annevk (~annevk@EM114-48-14-90.pool.e-mobile.ne.jp) (Ping timeout: 252 seconds)</li>
 <li id="l-189"><a href="#l-189">#</a> [12:42] * Quits: MikeSmith (~MikeSmith@EM114-48-14-90.pool.e-mobile.ne.jp) (Ping timeout: 258 seconds)</li>
 <li id="l-190" _a="zcorpan" _r="foolip"><a href="#l-190">#</a> [12:44] &lt;zcorpan&gt; foolip: i'm thinking either tolerating whitespace between cues or starting a new cue when seeing a timing line</li>
 <li id="l-191"><a href="#l-191">#</a> [12:44] * Quits: Rich_Clark (~chatzilla@94-193-82-82.zone7.bethere.co.uk) (Read error: Connection reset by peer)</li>
 <li id="l-192" _a="smaug____" _r="karlcow"><a href="#l-192">#</a> [12:45] &lt;smaug____&gt; karlcow: looks like they want to bring not only one but many proprietary stacks to web</li>
 <li id="l-193" _a="karlcow"><a href="#l-193">#</a> [12:46] &lt;karlcow&gt; yup looks like this</li>
 <li id="l-194" _a="zcorpan" _r="foolip"><a href="#l-194">#</a> [12:46] &lt;zcorpan&gt; foolip: where timing line is a line containing --&gt;. if you want --&gt; in data, you'd need to escape it</li>
 <li id="l-195"><a href="#l-195">#</a> [12:46] * Joins: annevk (~annevk@EM111-191-136-185.pool.e-mobile.ne.jp)</li>
 <li id="l-196" _a="karlcow"><a href="#l-196">#</a> [12:46] &lt;karlcow&gt; or I didn't understand and it also seems like an interop nightmare</li>
 <li id="l-197" _a="karlcow"><a href="#l-197">#</a> [12:47] &lt;karlcow&gt; and also the HTTP layer is placed in a strange place in their diagram.</li>
 <li id="l-198" _a="smaug____"><a href="#l-198">#</a> [12:47] &lt;smaug____&gt; it certainly looks like interop nightmare</li>
 <li id="l-199" _a="zcorpan" _r="foolip"><a href="#l-199">#</a> [12:47] &lt;zcorpan&gt; foolip: another thing that might be common is to not have two blank lines in &quot;header&quot;, so the first cue gets eaten</li>
 <li id="l-200" _a="smaug____" _r="oh"><a href="#l-200">#</a> [12:47] &lt;smaug____&gt; oh, hmm, http is nowhere near network</li>
 <li id="l-201" _a="foolip" _r="zcorpan"><a href="#l-201">#</a> [12:48] &lt;foolip&gt; zcorpan, I would also prefer to have timing lines always start new cues and require escaping to allow &quot;--&gt;&quot; inside cues</li>
 <li id="l-202" _a="karlcow"><a href="#l-202">#</a> [12:49] &lt;karlcow&gt; this post is a lot more fun <a href="http://sebleedelisle.com/2011/09/pixelphones-a-huge-display-made-with-smart-phones/">http://sebleedelisle.com/2011/09/pixelphones-a-huge-display-made-with-smart-phones/</a></li>
 <li id="l-203" _a="karlcow"><a href="#l-203">#</a> [12:49] &lt;karlcow&gt; websocket inside</li>
 <li id="l-204" _a="smaug____"><a href="#l-204">#</a> [12:49] &lt;smaug____&gt; people in research centers can propose many silly things. I certainly did at my time :)</li>
 <li id="l-205" _a="foolip" _r="zcorpan"><a href="#l-205">#</a> [12:49] &lt;foolip&gt; zcorpan, second best would be to strip whitespace when collecting a line, but meh</li>
 <li id="l-206" _a="karlcow"><a href="#l-206">#</a> [12:50] &lt;karlcow&gt; using the light sequences for localizing the phones is genius</li>
 <li id="l-207"><a href="#l-207">#</a> [12:51] * Joins: MikeSmith (~MikeSmith@EM111-191-136-185.pool.e-mobile.ne.jp)</li>
 <li id="l-208" _a="zcorpan" _r="foolip"><a href="#l-208">#</a> [12:54] &lt;zcorpan&gt; foolip: WDYT about cue in header?</li>
 <li id="l-209" _a="foolip" _r="zcorpan"><a href="#l-209">#</a> [12:54] &lt;foolip&gt; zcorpan, you mean directly following WEBVTT?</li>
 <li id="l-210" _a="zcorpan"><a href="#l-210">#</a> [12:54] &lt;zcorpan&gt; yes</li>
 <li id="l-211" _a="foolip"><a href="#l-211">#</a> [12:55] &lt;foolip&gt; if we make a timing line always start a new cue I don't see why we couldn't make that work as well</li>
 <li id="l-212"><a href="#l-212">#</a> [12:55] * Quits: zdobersek (~zan@89.142.251.175) (Quit: Leaving.)</li>
 <li id="l-213" _a="annevk"><a href="#l-213">#</a> [12:55] &lt;annevk&gt; we want this because 1.3% contains an error currently?</li>
 <li id="l-214" _a="annevk"><a href="#l-214">#</a> [12:55] &lt;annevk&gt; of SRT files, no less...</li>
 <li id="l-215"><a href="#l-215">#</a> [12:55] * Joins: zdobersek (~zan@89.142.251.175)</li>
 <li id="l-216" _a="foolip" _r="however"><a href="#l-216">#</a> [12:55] &lt;foolip&gt; however, that particular error you are actually quite likely to notice, since it's the first cue</li>
 <li id="l-217" _a="annevk"><a href="#l-217">#</a> [12:55] &lt;annevk&gt; the current format seems much easier to read</li>
 <li id="l-218" _a="annevk"><a href="#l-218">#</a> [12:56] &lt;annevk&gt; and we have a validator of sorts now :)</li>
 <li id="l-219" _a="foolip" _r="annevk"><a href="#l-219">#</a> [12:56] &lt;foolip&gt; annevk, except a validator doesn't complain about cues in cues, right?</li>
 <li id="l-220" _a="zcorpan" _r="annevk"><a href="#l-220">#</a> [12:56] &lt;zcorpan&gt; annevk: i don't see any reason the error would be less common in VTT</li>
 <li id="l-221" _a="annevk" _r="foolip"><a href="#l-221">#</a> [12:57] &lt;annevk&gt; foolip, it could warn</li>
 <li id="l-222" _a="foolip" _r="annevk"><a href="#l-222">#</a> [12:57] &lt;foolip&gt; annevk, but it's valid per the syntax right now</li>
 <li id="l-223" _a="zcorpan" _r="annevk"><a href="#l-223">#</a> [12:57] &lt;zcorpan&gt; annevk: &quot;easy to read&quot; is irrelevant since we're not suggesting changing what valid files look like</li>
 <li id="l-224" _a="annevk" _r="foolip"><a href="#l-224">#</a> [12:58] &lt;annevk&gt; foolip, yeah making --&gt; in cues invalid might be a good idea</li>
 <li id="l-225" _a="zcorpan"><a href="#l-225">#</a> [12:59] &lt;zcorpan&gt; making it invalid and start a new cue might be an even better idea :)</li>
 <li id="l-226" _a="zcorpan"><a href="#l-226">#</a> [12:59] &lt;zcorpan&gt; since it's what the author intended and what the user wants</li>
 <li id="l-227"><a href="#l-227">#</a> [12:59] * Joins: mpt (~mpt@canonical/mpt)</li>
 <li id="l-228" _a="zcorpan"><a href="#l-228">#</a> [12:59] &lt;zcorpan&gt; and what SRT does</li>
 <li id="l-229" _a="annevk"><a href="#l-229">#</a> [13:00] &lt;annevk&gt; euh okay...</li>
 <li id="l-230" _a="annevk"><a href="#l-230">#</a> [13:03] &lt;annevk&gt; SRT also does , and not .</li>
 <li id="l-231" _a="annevk"><a href="#l-231">#</a> [13:03] &lt;annevk&gt; still not sure why we don't do SRT if that's what everyone wants</li>
 <li id="l-232" _a="annevk"><a href="#l-232">#</a> [13:03] &lt;annevk&gt; can't really both argue we need a new format but UAs are also going to support SRT</li>
 <li id="l-233" _a="zcorpan"><a href="#l-233">#</a> [13:06] &lt;zcorpan&gt; SRT parsers support both , and .</li>
 <li id="l-234" _a="zcorpan"><a href="#l-234">#</a> [13:07] &lt;zcorpan&gt; personally i think we should have specified and extended SRT, but that's not where we're at</li>
 <li id="l-235" _a="zcorpan" _r="anyway"><a href="#l-235">#</a> [13:08] &lt;zcorpan&gt; anyway, i'm not doing this to end up with SRT. i'm doing this to get error handling in VTT better based on what errors people do in SRT</li>
 <li id="l-236" _a="zcorpan"><a href="#l-236">#</a> [13:09] &lt;zcorpan&gt; e.g. it turned out that errors in timestamps was pretty rare, so i'm not proposing to make that more lax after researching that</li>
 <li id="l-237"><a href="#l-237">#</a> [13:10] * Quits: MikeSmith (~MikeSmith@EM111-191-136-185.pool.e-mobile.ne.jp) (Ping timeout: 252 seconds)</li>
 <li id="l-238"><a href="#l-238">#</a> [13:11] * Quits: annevk (~annevk@EM111-191-136-185.pool.e-mobile.ne.jp) (Ping timeout: 258 seconds)</li>
 <li id="l-239"><a href="#l-239">#</a> [13:16] * Joins: necolas (~necolas@5e041dd3.bb.sky.com)</li>
 <li id="l-240"><a href="#l-240">#</a> [13:18] * Quits: smaug____ (~chatzilla@GYGMMCLXXXVI.gprs.sl-laajakaista.fi) (Ping timeout: 260 seconds)</li>
 <li id="l-241"><a href="#l-241">#</a> [13:18] * Quits: virtuelv (~virtuelv_@pat-tdc.opera.com) (Quit: Ex-Chat)</li>
 <li id="l-242"><a href="#l-242">#</a> [13:19] * Joins: Workshiva (~Dashiva@74.125.57.33)</li>
 <li id="l-243"><a href="#l-243">#</a> [13:21] * Joins: MikeSmith (~MikeSmith@EM114-48-118-80.pool.e-mobile.ne.jp)</li>
 <li id="l-244"><a href="#l-244">#</a> [13:21] * Quits: mpt (~mpt@canonical/mpt) (Ping timeout: 240 seconds)</li>
 <li id="l-245" _a="bga_"><a href="#l-245">#</a> [13:21] &lt;bga_&gt; <a href="http://kizu.github.com/Castle-of-Sinister-Sorcerers/">http://kizu.github.com/Castle-of-Sinister-Sorcerers/</a></li>
 <li id="l-246"><a href="#l-246">#</a> [13:31] * Joins: annevk (~annevk@EM114-48-118-80.pool.e-mobile.ne.jp)</li>
 <li id="l-247"><a href="#l-247">#</a> [13:40] * Joins: virtuelv (~virtuelv_@pat-tdc.opera.com)</li>
 <li id="l-248" _a="MikeSmith"><a href="#l-248">#</a> [13:42] &lt;MikeSmith&gt; anybody know when the content model for the &lt;dt&gt; element was changed to allow flow content?</li>
 <li id="l-249" _a="MikeSmith"><a href="#l-249">#</a> [13:42] &lt;MikeSmith&gt; (instead of just phrasing content)</li>
 <li id="l-250" _a="MikeSmith"><a href="#l-250">#</a> [13:44] &lt;MikeSmith&gt; seems it was some time after May 25</li>
 <li id="l-251" _a="zcorpan"><a href="#l-251">#</a> [13:48] &lt;zcorpan&gt; <a href="http://html5.org/tools/web-apps-tracker?from=6617&amp;to=6618">http://html5.org/tools/web-apps-tracker?from=6617&amp;to=6618</a></li>
 <li id="l-252"><a href="#l-252">#</a> [13:49] * Joins: jdong_bot_ (~jdong_bot@118.186.206.219)</li>
 <li id="l-253" _a="zcorpan"><a href="#l-253">#</a> [13:51] &lt;zcorpan&gt; hmm. &lt;b\n&gt; gets dropped but &lt;b \n&gt; is parsed</li>
 <li id="l-254"><a href="#l-254">#</a> [13:52] * Joins: _bga (~bga@ppp78-37-214-230.pppoe.avangarddsl.ru)</li>
 <li id="l-255" _a="annevk"><a href="#l-255">#</a> [13:52] &lt;annevk&gt; yeah white space rules in the parser are funky</li>
 <li id="l-256" _a="annevk"><a href="#l-256">#</a> [13:53] &lt;annevk&gt; I filed a bug on that</li>
 <li id="l-257" _a="annevk"><a href="#l-257">#</a> [13:53] &lt;annevk&gt; as they do not handle FF either</li>
 <li id="l-258" _a="annevk"><a href="#l-258">#</a> [13:53] &lt;annevk&gt; while the timing/setting algorithm does</li>
 <li id="l-259"><a href="#l-259">#</a> [13:54] * Quits: bga_ (~bga@ppp78-37-199-12.pppoe.avangarddsl.ru) (Ping timeout: 260 seconds)</li>
 <li id="l-260" _a="zcorpan"><a href="#l-260">#</a> [13:55] &lt;zcorpan&gt; yeah</li>
 <li id="l-261"><a href="#l-261">#</a> [13:57] * Quits: annevk (~annevk@EM114-48-118-80.pool.e-mobile.ne.jp) (Ping timeout: 252 seconds)</li>
 <li id="l-262"><a href="#l-262">#</a> [13:57] * Quits: MikeSmith (~MikeSmith@EM114-48-118-80.pool.e-mobile.ne.jp) (Ping timeout: 252 seconds)</li>
 <li id="l-263"><a href="#l-263">#</a> [13:59] * Quits: zdobersek (~zan@89.142.251.175) (Ping timeout: 252 seconds)</li>
 <li id="l-264"><a href="#l-264">#</a> [14:02] * Joins: annevk (~annevk@EM114-49-131-100.pool.e-mobile.ne.jp)</li>
 <li id="l-265"><a href="#l-265">#</a> [14:04] * Joins: MikeSmith (~MikeSmith@EM114-49-131-100.pool.e-mobile.ne.jp)</li>
 <li id="l-266" _a="MikeSmith" _r="zcorpan"><a href="#l-266">#</a> [14:06] &lt;MikeSmith&gt; zcorpan: thanks</li>
 <li id="l-267"><a href="#l-267">#</a> [14:15] * Joins: benjoffe_ (~benjoffe_@CPE-121-218-141-115.lnse3.cht.bigpond.net.au)</li>
 <li id="l-268"><a href="#l-268">#</a> [14:22] * Joins: mpt (mpt@conference/ubuntudevelopersummit/x-donwfckmoiygtras)</li>
 <li id="l-269"><a href="#l-269">#</a> [14:22] * Quits: mpt (mpt@conference/ubuntudevelopersummit/x-donwfckmoiygtras) (Changing host)</li>
 <li id="l-270"><a href="#l-270">#</a> [14:22] * Joins: mpt (mpt@canonical/mpt)</li>
 <li id="l-271"><a href="#l-271">#</a> [14:26] * Joins: nimbupani (~divyam@219.64.117.145)</li>
 <li id="l-272" _a="nimbupani"><a href="#l-272">#</a> [14:26] * nimbupani is now known as divya</li>
 <li id="l-273"><a href="#l-273">#</a> [14:29] * Quits: annevk (~annevk@EM114-49-131-100.pool.e-mobile.ne.jp) (Ping timeout: 252 seconds)</li>
 <li id="l-274"><a href="#l-274">#</a> [14:31] * Quits: MikeSmith (~MikeSmith@EM114-49-131-100.pool.e-mobile.ne.jp) (Ping timeout: 276 seconds)</li>
 <li id="l-275"><a href="#l-275">#</a> [14:33] * Quits: jdong_bot_ (~jdong_bot@118.186.206.219) (Remote host closed the connection)</li>
 <li id="l-276"><a href="#l-276">#</a> [14:38] * Joins: ezoe (~ezoe@112-68-245-190f1.kyt1.eonet.ne.jp)</li>
 <li id="l-277"><a href="#l-277">#</a> [14:42] * Joins: MikeSmith (~MikeSmith@EM1-113-75-77.pool.e-mobile.ne.jp)</li>
 <li id="l-278"><a href="#l-278">#</a> [14:50] * Quits: lhnz (~lhnz@188-223-83-48.zone14.bethere.co.uk) (Ping timeout: 252 seconds)</li>
 <li id="l-279"><a href="#l-279">#</a> [14:51] * Joins: lhnz (~lhnz@188-223-83-48.zone14.bethere.co.uk)</li>
 <li id="l-280" _a="zcorpan" _r="MikeSmith"><a href="#l-280">#</a> [14:55] &lt;zcorpan&gt; MikeSmith: why was <a href="http://www.w3.org/Bugs/Public/show_bug.cgi?id=14552">http://www.w3.org/Bugs/Public/show_bug.cgi?id=14552</a> not emailed to public-html?</li>
 <li id="l-281" _a="MikeSmith" _r="zcorpan"><a href="#l-281">#</a> [14:56] &lt;MikeSmith&gt; zcorpan: maybe because of teh component</li>
 <li id="l-282" _a="MikeSmith"><a href="#l-282">#</a> [14:56] &lt;MikeSmith&gt; checking now</li>
 <li id="l-283" _a="MikeSmith"><a href="#l-283">#</a> [14:57] &lt;MikeSmith&gt; yeah</li>
 <li id="l-284" _a="MikeSmith"><a href="#l-284">#</a> [14:57] &lt;MikeSmith&gt; I have &quot;other Hixie drafts&quot; component set up to don't generate notifications to public-html</li>
 <li id="l-285" _a="MikeSmith"><a href="#l-285">#</a> [14:58] &lt;MikeSmith&gt; but I can change that if you wan</li>
 <li id="l-286" _a="zcorpan" _r="foolip"><a href="#l-286">#</a> [14:59] &lt;zcorpan&gt; foolip: ^</li>
 <li id="l-287"><a href="#l-287">#</a> [14:59] * Quits: virtuelv (~virtuelv_@pat-tdc.opera.com) (Quit: Ex-Chat)</li>
 <li id="l-288"><a href="#l-288">#</a> [15:02] * Joins: temp02 (~temp01@unaffiliated/temp01)</li>
 <li id="l-289"><a href="#l-289">#</a> [15:04] * Joins: zdobersek (~zan@89.142.251.175)</li>
 <li id="l-290"><a href="#l-290">#</a> [15:05] * Quits: temp01 (~temp01@unaffiliated/temp01) (Disconnected by services)</li>
 <li id="l-291" _a="temp02"><a href="#l-291">#</a> [15:05] * temp02 is now known as temp01</li>
 <li id="l-292"><a href="#l-292">#</a> [15:07] * Joins: GlitchMr (~glitchmr@178-36-55-53.adsl.inetia.pl)</li>
 <li id="l-293"><a href="#l-293">#</a> [15:07] * Quits: akamike (~akamike@94-193-106-14.zone7.bethere.co.uk) (Quit: akamike)</li>
 <li id="l-294"><a href="#l-294">#</a> [15:08] * Quits: MikeSmith (~MikeSmith@EM1-113-75-77.pool.e-mobile.ne.jp) (Quit: MikeSmith)</li>
 <li id="l-295"><a href="#l-295">#</a> [15:09] * Joins: davidb_ (~davidb@66.207.208.98)</li>
 <li id="l-296"><a href="#l-296">#</a> [15:10] * Quits: yuuki_ (~kobayashi@58x158x182x50.ap58.ftth.ucom.ne.jp) (Quit: Leaving...)</li>
 <li id="l-297"><a href="#l-297">#</a> [15:12] * Joins: MacTed (~Thud@63.119.36.36)</li>
 <li id="l-298" _a="foolip"><a href="#l-298">#</a> [15:15] &lt;foolip&gt; MikeSmith (when back), was this changed recently? I've filed loads of bugs and can't remember this happening before</li>
 <li id="l-299"><a href="#l-299">#</a> [15:21] * Quits: tomasf (~tomasf@77.72.97.5.c.fiberdirekt.net) (Quit: tomasf)</li>
 <li id="l-300"><a href="#l-300">#</a> [15:24] * Joins: miketaylr (~miketaylr@206.217.92.186)</li>
 <li id="l-301"><a href="#l-301">#</a> [15:43] * Quits: Areks (~Areks@rs.gridnine.com) (Read error: Connection reset by peer)</li>
 <li id="l-302"><a href="#l-302">#</a> [15:51] * Joins: MikeSmith (~MikeSmith@EM1-113-75-77.pool.e-mobile.ne.jp)</li>
 <li id="l-303"><a href="#l-303">#</a> [15:54] * Quits: GlitchMr (~glitchmr@178-36-55-53.adsl.inetia.pl) (Read error: Connection reset by peer)</li>
 <li id="l-304"><a href="#l-304">#</a> [15:55] * Quits: zdobersek (~zan@89.142.251.175) (Read error: Connection reset by peer)</li>
 <li id="l-305"><a href="#l-305">#</a> [15:55] * Joins: zdobersek (~zan@89.142.251.175)</li>
 <li id="l-306"><a href="#l-306">#</a> [16:01] * Joins: scor (~scor@drupal.org/user/52142/view)</li>
 <li id="l-307"><a href="#l-307">#</a> [16:10] * Quits: agektmr (~Adium@2401:fa00:4:1012:fa1e:dfff:fee6:d74e) (Quit: Leaving.)</li>
 <li id="l-308"><a href="#l-308">#</a> [16:11] * Joins: rillian_ (~rillian@mist.thaumas.net)</li>
 <li id="l-309"><a href="#l-309">#</a> [16:12] * Joins: annevk (~annevk@EM1-113-75-77.pool.e-mobile.ne.jp)</li>
 <li id="l-310" _a="annevk" _r="foolip"><a href="#l-310">#</a> [16:13] &lt;annevk&gt; foolip, &lt;track&gt; is in W3C HTML, WebVTT is not</li>
 <li id="l-311" _a="annevk" _r="foolip"><a href="#l-311">#</a> [16:13] &lt;annevk&gt; foolip, I filed quite a few bugs on WebVTT nobody got emailed about either :(</li>
 <li id="l-312" _a="foolip" _r="annevk"><a href="#l-312">#</a> [16:18] &lt;foolip&gt; annevk, how can I find them?</li>
 <li id="l-313" _a="annevk"><a href="#l-313">#</a> [16:19] &lt;annevk&gt; 14536 - 14543</li>
 <li id="l-314" _a="annevk"><a href="#l-314">#</a> [16:19] &lt;annevk&gt; and 14547</li>
 <li id="l-315" _a="foolip"><a href="#l-315">#</a> [16:19] &lt;foolip&gt; will add &lt;track&gt; to all of them</li>
 <li id="l-316" _a="annevk"><a href="#l-316">#</a> [16:19] &lt;annevk&gt; thanks!</li>
 <li id="l-317" _a="rillian_"><a href="#l-317">#</a> [16:20] &lt;rillian_&gt; indeed!</li>
 <li id="l-318" _a="annevk"><a href="#l-318">#</a> [16:20] &lt;annevk&gt; I will add &lt;track&gt; in the future</li>
 <li id="l-319" _a="rillian_"><a href="#l-319">#</a> [16:20] &lt;rillian_&gt; thanks annevk, I didn't see them either</li>
 <li id="l-320"><a href="#l-320">#</a> [16:23] * Joins: tomasf (~tomasf@host-95-199-3-156.mobileonline.telia.com)</li>
 <li id="l-321" _a="foolip" _r="annevk"><a href="#l-321">#</a> [16:23] &lt;foolip&gt; annevk, check <a href="http://wiki.whatwg.org/wiki/WebVTT">http://wiki.whatwg.org/wiki/WebVTT</a> to see if they're all listed</li>
 <li id="l-322" _a="foolip"><a href="#l-322">#</a> [16:23] &lt;foolip&gt; will mark the non-editorial ones as blocker as well</li>
 <li id="l-323" _a="foolip" _r="annevk"><a href="#l-323">#</a> [16:33] &lt;foolip&gt; annevk, <a href="http://www.w3.org/Bugs/Public/show_bug.cgi?id=14547">http://www.w3.org/Bugs/Public/show_bug.cgi?id=14547</a> is not very clear</li>
 <li id="l-324" _a="foolip"><a href="#l-324">#</a> [16:34] &lt;foolip&gt; two issues in one</li>
 <li id="l-325" _a="foolip" _r="oops"><a href="#l-325">#</a> [16:34] &lt;foolip&gt; oops, was I mistaken to mark <a href="http://www.w3.org/Bugs/Public/show_bug.cgi?id=14551">http://www.w3.org/Bugs/Public/show_bug.cgi?id=14551</a> as &lt;track&gt;?</li>
 <li id="l-326" _a="zcorpan"><a href="#l-326">#</a> [16:36] &lt;zcorpan&gt; you didn't mark it as &lt;track&gt; :)</li>
 <li id="l-327" _a="foolip" _r="zcorpan"><a href="#l-327">#</a> [16:36] &lt;foolip&gt; zcorpan, did you?</li>
 <li id="l-328" _a="foolip" _r="oh"><a href="#l-328">#</a> [16:36] &lt;foolip&gt; oh, I'm not even CC'd, it must have been someone else</li>
 <li id="l-329" _a="zcorpan"><a href="#l-329">#</a> [16:37] &lt;zcorpan&gt; i filed it</li>
 <li id="l-330" _a="annevk" _r="foolip"><a href="#l-330">#</a> [16:37] &lt;annevk&gt; foolip, tried to clarify it, did not bother with splitting the issue for now...</li>
 <li id="l-331" _a="AryehGregor" _r="jgraham"><a href="#l-331">#</a> [16:39] &lt;AryehGregor&gt; jgraham, 1) If we're supposed to still use testharnessreport.js, why does the &quot;Basic usage&quot; section in testharness.js docs at the top of the file not mention it?  It just says to add &lt;script src=&quot;<a href="http://w3c-test.org/resources/testharness.js&quot;&gt;&lt;/script">http://w3c-test.org/resources/testharness.js&quot;&gt;&lt;/script</a>&gt;</li>
 <li id="l-332" _a="AryehGregor"><a href="#l-332">#</a> [16:39] &lt;AryehGregor&gt; And 2) why is using an absolute URL wrong?  Using a relative URL like that will break any checkout of the tests other than w3c-test.org.  Users of other checkouts might not be able to put a resources directory in the web root.</li>
 <li id="l-333" _a="AryehGregor"><a href="#l-333">#</a> [16:40] &lt;AryehGregor&gt; Or even if they can, why should they have to?</li>
 <li id="l-334" _a="AryehGregor"><a href="#l-334">#</a> [16:42] &lt;AryehGregor&gt; I guess using a relative URL makes sense if you want to test what effect changes to the harness have on webapps tests, so you don't have to manually change the &lt;script&gt; URLs.</li>
 <li id="l-335" _a="AryehGregor"><a href="#l-335">#</a> [16:42] &lt;AryehGregor&gt; (which is what I was doing)</li>
 <li id="l-336" _a="AryehGregor" _r="Anyway"><a href="#l-336">#</a> [16:42] &lt;AryehGregor&gt; Anyway, if we want people to use relative URLs, testharness.js should say that.  The way it says it should be used is with an absolute URL.</li>
 <li id="l-337" _a="AryehGregor" _r="annevk"><a href="#l-337">#</a> [16:42] &lt;AryehGregor&gt; annevk, ^^</li>
 <li id="l-338"><a href="#l-338">#</a> [16:44] * Joins: spliter_ (~spliter@164.247.189.109.customer.cdi.no)</li>
 <li id="l-339" _a="jgraham" _r="AryehGregor"><a href="#l-339">#</a> [16:45] &lt;jgraham&gt; AryehGregor: That example should prbably be changed</li>
 <li id="l-340" _a="AryehGregor" _r="jgraham"><a href="#l-340">#</a> [16:45] &lt;AryehGregor&gt; jgraham, please do.  It's what I was looking out to figure out how the harness should be used.</li>
 <li id="l-341" _a="jgraham"><a href="#l-341">#</a> [16:46] &lt;jgraham&gt; Using absolute urls is wrong because it makes it harder for people to import the tests into their local testing systems</li>
 <li id="l-342" _a="timeless"><a href="#l-342">#</a> [16:47] &lt;timeless&gt; i'm not sure that's really true fwiw</li>
 <li id="l-343" _a="jgraham"><a href="#l-343">#</a> [16:47] &lt;jgraham&gt; Using resources on external servers is horrible badness so one would need to muck about with DNS or something</li>
 <li id="l-344" _a="timeless"><a href="#l-344">#</a> [16:47] &lt;timeless&gt; mozilla's testharness actually redirects dns lookups iirc</li>
 <li id="l-345" _a="AryehGregor"><a href="#l-345">#</a> [16:47] &lt;AryehGregor&gt; Or you could just do a search-and-replace on all the URLs as part of the process of checking them out.</li>
 <li id="l-346" _a="AryehGregor"><a href="#l-346">#</a> [16:47] &lt;AryehGregor&gt; That would work too.</li>
 <li id="l-347" _a="timeless"><a href="#l-347">#</a> [16:47] &lt;timeless&gt; search and replace otoh tends to lead to merge hell</li>
 <li id="l-348" _a="jgraham" _r="AryehGregor"><a href="#l-348">#</a> [16:47] &lt;jgraham&gt; AryehGregor: Non-trivial patches are always a source of pain</li>
 <li id="l-349" _a="AryehGregor"><a href="#l-349">#</a> [16:47] &lt;AryehGregor&gt; You can search-and-replace it back before doing any kind of merge.</li>
 <li id="l-350" _a="zewt"><a href="#l-350">#</a> [16:48] &lt;zewt&gt; D:</li>
 <li id="l-351" _a="jgraham" _r="AryehGregor"><a href="#l-351">#</a> [16:48] &lt;jgraham&gt; AryehGregor: GLWT if the person who made the original changes isn't the same person doing the update</li>
 <li id="l-352" _a="jgraham" _r="timeless"><a href="#l-352">#</a> [16:48] &lt;jgraham&gt; timeless: Yes, I think Mozilla do that. We typically don't (but could of course)</li>
 <li id="l-353" _a="AryehGregor"><a href="#l-353">#</a> [16:49] &lt;AryehGregor&gt; If we're going to do it this way, then 1) don't assume people have control of the web root, or want to add a resources directory there -- make it relative to the current directory, not the root.  2) Post clear instructions somewhere on how you're supposed to clone the webapps tests, because as it stands a simple clone will be completely broken.</li>
 <li id="l-354" _a="jgraham"><a href="#l-354">#</a> [16:49] &lt;jgraham&gt; In any case using relative URLs isn't bad for browser vendors, but absolue urls can be</li>
 <li id="l-355" _a="timeless" _r="fwiw"><a href="#l-355">#</a> [16:49] &lt;timeless&gt; fwiw, i'm not really sure where i sit on this</li>
 <li id="l-356" _a="timeless"><a href="#l-356">#</a> [16:49] &lt;timeless&gt; i have a spec i'm working on which has the absolute url in it</li>
 <li id="l-357" _a="AryehGregor" _r="Also"><a href="#l-357">#</a> [16:49] &lt;AryehGregor&gt; Also, the dvcs.w3.org versions should really work if possible, so that you can easily look at old versions of the tests.</li>
 <li id="l-358" _a="timeless"><a href="#l-358">#</a> [16:50] &lt;timeless&gt; and i kind of wanted to be able to have a local instance w/o web access and w/o having to install a test harness</li>
 <li id="l-359" _a="jgraham" _r="AryehGregor"><a href="#l-359">#</a> [16:50] &lt;jgraham&gt; AryehGregor: I think for any non-trivial use you will have to install locally</li>
 <li id="l-360" _a="AryehGregor"><a href="#l-360">#</a> [16:50] &lt;AryehGregor&gt; Plus because it's confusing that they don't work.</li>
 <li id="l-361" _a="AryehGregor" _r="jgraham"><a href="#l-361">#</a> [16:50] &lt;AryehGregor&gt; jgraham, yes, if only so that your tests don't suddenly behave differently one day because the remote copy changed . . .</li>
 <li id="l-362" _a="AryehGregor" _r="jgraham"><a href="#l-362">#</a> [16:55] &lt;AryehGregor&gt; jgraham, is there a place to file bugs on the harness?</li>
 <li id="l-363"><a href="#l-363">#</a> [16:55] * Quits: zdobersek (~zan@89.142.251.175) (Quit: Leaving.)</li>
 <li id="l-364"><a href="#l-364">#</a> [16:55] * Joins: zdobersek (~zan@89.142.251.175)</li>
 <li id="l-365"><a href="#l-365">#</a> [16:56] * Joins: davidwalsh (~davidwals@75-135-74-55.dhcp.mdsn.wi.charter.com)</li>
 <li id="l-366" _a="jgraham" _r="AryehGregor"><a href="#l-366">#</a> [16:56] &lt;jgraham&gt; AryehGregor: I think bugzilla has a componen</li>
 <li id="l-367" _a="jgraham"><a href="#l-367">#</a> [16:56] &lt;jgraham&gt; t</li>
 <li id="l-368" _a="jgraham"><a href="#l-368">#</a> [16:56] &lt;jgraham&gt; Dunno if I get email though</li>
 <li id="l-369" _a="jgraham"><a href="#l-369">#</a> [16:56] &lt;jgraham&gt; I hink I might</li>
 <li id="l-370" _a="jgraham"><a href="#l-370">#</a> [16:57] &lt;jgraham&gt; +t</li>
 <li id="l-371" _a="jgraham"><a href="#l-371">#</a> [16:57] &lt;jgraham&gt; Anyway fixed the documentation</li>
 <li id="l-372" _a="jgraham"><a href="#l-372">#</a> [16:58] &lt;jgraham&gt; I think you could fix the webapps wiki to tell people how to set up a local copy of the testsuite</li>
 <li id="l-373" _a="jgraham"><a href="#l-373">#</a> [16:58] &lt;jgraham&gt; Really testharness.js should be in its own repository</li>
 <li id="l-374" _a="jgraham"><a href="#l-374">#</a> [16:58] &lt;jgraham&gt; Or one called &quot;resources&quot;</li>
 <li id="l-375"><a href="#l-375">#</a> [17:02] * Joins: KillerX (~anant@nat/mozilla/x-yftvkpxschruugkb)</li>
 <li id="l-376" _a="AryehGregor"><a href="#l-376">#</a> [17:03] &lt;AryehGregor&gt; Is there any way to set up hg so that it automatically clones a second repo when you clone the first?</li>
 <li id="l-377" _a="AryehGregor"><a href="#l-377">#</a> [17:03] &lt;AryehGregor&gt; So cloning the tests also clones the harness?</li>
 <li id="l-378" _a="AryehGregor"><a href="#l-378">#</a> [17:03] &lt;AryehGregor&gt; Like SVN deps or git submodules or whatever?</li>
 <li id="l-379" _a="timeless"><a href="#l-379">#</a> [17:03] &lt;timeless&gt; hg subrepo</li>
 <li id="l-380"><a href="#l-380">#</a> [17:03] * Joins: kennyluck (~kennyluck@114-43-119-208.dynamic.hinet.net)</li>
 <li id="l-381" _a="AryehGregor"><a href="#l-381">#</a> [17:03] * AryehGregor hasn't used either of those features much, so might not understand how they work</li>
 <li id="l-382" _a="timeless"><a href="#l-382">#</a> [17:04] &lt;timeless&gt; basically you create a repo and add a .hgsub file which lists the two other repos you actually want</li>
 <li id="l-383" _a="AryehGregor"><a href="#l-383">#</a> [17:04] &lt;AryehGregor&gt; So it would still be possible to clone the subrepo separately even though that might make no sense, I guess?</li>
 <li id="l-384" _a="timeless"><a href="#l-384">#</a> [17:05] &lt;timeless&gt; yes</li>
 <li id="l-385" _a="timeless"><a href="#l-385">#</a> [17:05] &lt;timeless&gt; <a href="http://stackoverflow.com/questions/5621670/hg-sub-repository-dependencies">http://stackoverflow.com/questions/5621670/hg-sub-repository-dependencies</a></li>
 <li id="l-386" _a="timeless"><a href="#l-386">#</a> [17:05] &lt;timeless&gt; is probably worth reading (i haven't)</li>
 <li id="l-387" _a="timeless"><a href="#l-387">#</a> [17:05] &lt;timeless&gt; you could skip the mostly empty repo if you like and just have the main content one also have the subrepo directives</li>
 <li id="l-388" _a="timeless"><a href="#l-388">#</a> [17:05] &lt;timeless&gt; but if someone for instance has 5 wgs</li>
 <li id="l-389" _a="timeless"><a href="#l-389">#</a> [17:05] &lt;timeless&gt; or specs</li>
 <li id="l-390" _a="AryehGregor"><a href="#l-390">#</a> [17:05] * AryehGregor is catching up on ten days of e-mail</li>
 <li id="l-391" _a="timeless"><a href="#l-391">#</a> [17:06] &lt;timeless&gt; and each spec has a sub dependency for the same third party repo</li>
 <li id="l-392" _a="timeless"><a href="#l-392">#</a> [17:06] &lt;timeless&gt; (call it &quot;testharness&quot;)</li>
 <li id="l-393" _a="timeless"><a href="#l-393">#</a> [17:06] &lt;timeless&gt; then making them have 5 copies as subrepos is cruel</li>
 <li id="l-394" _a="timeless"><a href="#l-394">#</a> [17:06] &lt;timeless&gt; when they could instead have 5 repos as siblings to their one copy of that testharness repo</li>
 <li id="l-395" _a="timeless"><a href="#l-395">#</a> [17:06] &lt;timeless&gt; make sense?</li>
 <li id="l-396" _a="timeless"><a href="#l-396">#</a> [17:07] &lt;timeless&gt; <a href="http://mercurial.selenic.com/wiki/Subrepository">http://mercurial.selenic.com/wiki/Subrepository</a> is the official documentation</li>
 <li id="l-397"><a href="#l-397">#</a> [17:11] * Joins: rillian__ (~rillian@mist.thaumas.net)</li>
 <li id="l-398"><a href="#l-398">#</a> [17:11] * Quits: rillian_ (~rillian@mist.thaumas.net) (Read error: Connection reset by peer)</li>
 <li id="l-399"><a href="#l-399">#</a> [17:11] * Joins: rabbi1 (~manjunath@49.249.126.64)</li>
 <li id="l-400"><a href="#l-400">#</a> [17:14] * Joins: rillian_ (~rillian@mist.thaumas.net)</li>
 <li id="l-401"><a href="#l-401">#</a> [17:14] * Quits: rillian__ (~rillian@mist.thaumas.net) (Read error: Connection reset by peer)</li>
 <li id="l-402"><a href="#l-402">#</a> [17:14] * Joins: rillian__ (~rillian@mist.thaumas.net)</li>
 <li id="l-403"><a href="#l-403">#</a> [17:14] * Quits: rillian_ (~rillian@mist.thaumas.net) (Read error: Connection reset by peer)</li>
 <li id="l-404"><a href="#l-404">#</a> [17:15] * Quits: rillian__ (~rillian@mist.thaumas.net) (Read error: Connection reset by peer)</li>
 <li id="l-405"><a href="#l-405">#</a> [17:16] * Joins: rillian_ (~rillian@mist.thaumas.net)</li>
 <li id="l-406"><a href="#l-406">#</a> [17:17] * Quits: davidwalsh (~davidwals@75-135-74-55.dhcp.mdsn.wi.charter.com) (Quit: Reading <a href="http://davidwalsh.name">http://davidwalsh.name</a>)</li>
 <li id="l-407"><a href="#l-407">#</a> [17:17] * Quits: zcorpan (~zcorpan@c-699de355.410-6-64736c14.cust.bredbandsbolaget.se) (Quit: zcorpan)</li>
 <li id="l-408"><a href="#l-408">#</a> [17:21] * Quits: necolas (~necolas@5e041dd3.bb.sky.com) (Remote host closed the connection)</li>
 <li id="l-409"><a href="#l-409">#</a> [17:21] * Quits: spliter_ (~spliter@164.247.189.109.customer.cdi.no) (Quit: spliter_)</li>
 <li id="l-410"><a href="#l-410">#</a> [17:22] * Joins: temp02 (~temp01@unaffiliated/temp01)</li>
 <li id="l-411"><a href="#l-411">#</a> [17:23] * Quits: Evanescence (~Evanescen@122.237.28.26) (Quit: my website: <a href="http://stardiviner.dyndns-blog.com/">http://stardiviner.dyndns-blog.com/</a>)</li>
 <li id="l-412"><a href="#l-412">#</a> [17:23] * Quits: temp01 (~temp01@unaffiliated/temp01) (Ping timeout: 240 seconds)</li>
 <li id="l-413" _a="_bga"><a href="#l-413">#</a> [17:24] &lt;_bga&gt; oh new js min trend, if(a != 15) -&gt; if(a^15)</li>
 <li id="l-414"><a href="#l-414">#</a> [17:25] * Joins: shepazu (~shepazu@108-70-132-46.lightspeed.rlghnc.sbcglobal.net)</li>
 <li id="l-415"><a href="#l-415">#</a> [17:25] * Joins: AlexNRoss (~AleossIRC@unaffiliated/aleoss)</li>
 <li id="l-416" _a="nlogax"><a href="#l-416">#</a> [17:27] &lt;nlogax&gt; too bad they give different results</li>
 <li id="l-417" _a="_bga"><a href="#l-417">#</a> [17:28] &lt;_bga&gt; for a = 15.3 - yes</li>
 <li id="l-418" _a="nlogax"><a href="#l-418">#</a> [17:28] &lt;nlogax&gt; 4294967311 ^ 15, 4294967311 != 15, oops</li>
 <li id="l-419"><a href="#l-419">#</a> [17:28] * Joins: mokush (~quassel@188.24.43.163)</li>
 <li id="l-420" _a="_bga"><a href="#l-420">#</a> [17:28] &lt;_bga&gt; too</li>
 <li id="l-421" _a="nlogax"><a href="#l-421">#</a> [17:29] &lt;nlogax&gt; but i'm sure it will take off anyway :D</li>
 <li id="l-422"><a href="#l-422">#</a> [17:29] * Quits: temp02 (~temp01@unaffiliated/temp01) (Ping timeout: 258 seconds)</li>
 <li id="l-423"><a href="#l-423">#</a> [17:29] * Joins: temp02 (~temp01@unaffiliated/temp01)</li>
 <li id="l-424"><a href="#l-424">#</a> [17:31] * Quits: mokush (~quassel@188.24.43.163) (Remote host closed the connection)</li>
 <li id="l-425"><a href="#l-425">#</a> [17:34] * Joins: tndH (~Rob@cpc16-seac19-2-0-cust234.7-2.cable.virginmedia.com)</li>
 <li id="l-426" _a="_bga"><a href="#l-426">#</a> [17:37] &lt;_bga&gt; hm. if(a-15) !</li>
 <li id="l-427" _a="temp02"><a href="#l-427">#</a> [17:38] * temp02 is now known as temp01</li>
 <li id="l-428"><a href="#l-428">#</a> [17:39] * Joins: GlitchMr (~glitchmr@178-36-55-53.adsl.inetia.pl)</li>
 <li id="l-429" _a="zewt"><a href="#l-429">#</a> [17:39] &lt;zewt&gt; a somewhat more effective way of making code smaller: deflate.</li>
 <li id="l-430" _a="Philip`"><a href="#l-430">#</a> [17:39] &lt;Philip`&gt; An even more effective way: write less code</li>
 <li id="l-431" _a="_bga"><a href="#l-431">#</a> [17:40] &lt;_bga&gt; i know but for guys from 140bytes its sport</li>
 <li id="l-432"><a href="#l-432">#</a> [17:44] * Quits: tomasf (~tomasf@host-95-199-3-156.mobileonline.telia.com) (Quit: tomasf)</li>
 <li id="l-433" _a="dglazkov|away"><a href="#l-433">#</a> [17:46] * dglazkov|away is now known as dglazkov</li>
 <li id="l-434" class="flagged" _a="dglazkov"><a href="#l-434">#</a> [17:47] &lt;dglazkov&gt; good morning, Whatwg!</li>
 <li id="l-435" class="flagged" _a="divya"><a href="#l-435">#</a> [17:49] &lt;divya&gt; காலை வனக்கம் dglazkov!</li>
 <li id="l-436" _a="dglazkov"><a href="#l-436">#</a> [17:49] &lt;dglazkov&gt; that looks curly</li>
 <li id="l-437" _a="_bga"><a href="#l-437">#</a> [17:50] &lt;_bga&gt; i need google translate plugin for IRC client</li>
 <li id="l-438" _a="nlogax"><a href="#l-438">#</a> [17:51] &lt;nlogax&gt; <a href="http://scripts.irssi.org/scripts/gtrans.pl">http://scripts.irssi.org/scripts/gtrans.pl</a></li>
 <li id="l-439" _a="_bga"><a href="#l-439">#</a> [17:53] &lt;_bga&gt; thanks :)</li>
 <li id="l-440"><a href="#l-440">#</a> [17:53] * Quits: astearns (~anonymous@c-50-132-63-33.hsd1.wa.comcast.net) (Quit: astearns)</li>
 <li id="l-441"><a href="#l-441">#</a> [17:54] * Joins: Areks (~Areks@176.14.214.163)</li>
 <li id="l-442"><a href="#l-442">#</a> [17:55] * Joins: zdobersek1 (~zan@BSN-142-43-64.dial-up.dsl.siol.net)</li>
 <li id="l-443"><a href="#l-443">#</a> [17:55] * Quits: zdobersek (~zan@89.142.251.175) (Ping timeout: 244 seconds)</li>
 <li id="l-444"><a href="#l-444">#</a> [17:58] * Joins: mokush (~quassel@188.24.43.163)</li>
 <li id="l-445"><a href="#l-445">#</a> [17:58] * Quits: mpt (mpt@canonical/mpt) (Ping timeout: 240 seconds)</li>
 <li id="l-446"><a href="#l-446">#</a> [17:59] * Joins: mpt (mpt@canonical/mpt)</li>
 <li id="l-447" _a="annevk" _r="AryehGregor"><a href="#l-447">#</a> [18:00] &lt;annevk&gt; AryehGregor, I didn't know about the bogus documentation (does it also say to not have testharnessreport.js?), hopefully it's not too hard to revert</li>
 <li id="l-448" _a="hsivonen"><a href="#l-448">#</a> [18:00] &lt;hsivonen&gt; Gnome Terminal fails with Tamil :-(</li>
 <li id="l-449"><a href="#l-449">#</a> [18:03] * Joins: jdong_bot_ (~jdong_bot@118.186.129.129)</li>
 <li id="l-450"><a href="#l-450">#</a> [18:09] * Joins: Maurice (copyman@5ED573FA.cm-7-6b.dynamic.ziggo.nl)</li>
 <li id="l-451"><a href="#l-451">#</a> [18:09] * Joins: tomasf (~tom@c-5ed9e555.024-204-6c6b7012.cust.bredbandsbolaget.se)</li>
 <li id="l-452"><a href="#l-452">#</a> [18:11] * Quits: benjoffe_ (~benjoffe_@CPE-121-218-141-115.lnse3.cht.bigpond.net.au) (Read error: Connection reset by peer)</li>
 <li id="l-453" _a="dglazkov" _r="divya"><a href="#l-453">#</a> [18:15] &lt;dglazkov&gt; divya: what language is it?</li>
 <li id="l-454" _a="divya" _r="dglazkov"><a href="#l-454">#</a> [18:16] &lt;divya&gt; dglazkov: tamil :)</li>
 <li id="l-455" _a="divya"><a href="#l-455">#</a> [18:16] &lt;divya&gt; CURLYY</li>
 <li id="l-456" _a="divya"><a href="#l-456">#</a> [18:16] &lt;divya&gt; hahaha it is curly</li>
 <li id="l-457" _a="dglazkov"><a href="#l-457">#</a> [18:17] &lt;dglazkov&gt; cool. India is a crazy place. So many languages. So many people. So many foods to try</li>
 <li id="l-458"><a href="#l-458">#</a> [18:17] * Joins: benjoffe_ (~benjoffe_@CPE-121-218-141-115.lnse3.cht.bigpond.net.au)</li>
 <li id="l-459" _a="divya"><a href="#l-459">#</a> [18:17] &lt;divya&gt; <a href="http://translate.google.com/#auto|en">http://translate.google.com/#auto|en</a>|காலை%20வனக்கம்</li>
 <li id="l-460"><a href="#l-460">#</a> [18:17] * Quits: erlehmann (~erlehmann@dslb-084-062-029-250.pools.arcor-ip.net) (Quit: Ex-Chat)</li>
 <li id="l-461" _a="divya"><a href="#l-461">#</a> [18:17] &lt;divya&gt; almost there I meant good morning. :P</li>
 <li id="l-462" _a="divya"><a href="#l-462">#</a> [18:17] &lt;divya&gt; it is very crazy</li>
 <li id="l-463" _a="divya"><a href="#l-463">#</a> [18:17] &lt;divya&gt; and I am trying all the delish foood now.</li>
 <li id="l-464" _a="dglazkov"><a href="#l-464">#</a> [18:20] &lt;dglazkov&gt; has anyone figured out why some countries are really big on delicious food and some are like, &quot;Potatoes!!!&quot;</li>
 <li id="l-465" _a="divya"><a href="#l-465">#</a> [18:21] &lt;divya&gt; hahahah</li>
 <li id="l-466" _a="divya"><a href="#l-466">#</a> [18:21] &lt;divya&gt; good point.</li>
 <li id="l-467" _a="divya"><a href="#l-467">#</a> [18:22] &lt;divya&gt; ya funny thing chillies were not native to India but indians are big on spice.</li>
 <li id="l-468" _a="divya"><a href="#l-468">#</a> [18:22] &lt;divya&gt; err i mean heat</li>
 <li id="l-469" _a="heycam|away"><a href="#l-469">#</a> [18:22] * heycam|away is now known as heycam</li>
 <li id="l-470"><a href="#l-470">#</a> [18:23] * Quits: mokush (~quassel@188.24.43.163) (Remote host closed the connection)</li>
 <li id="l-471"><a href="#l-471">#</a> [18:24] * Quits: nielsle_ (~nielsle@3239059-cl69.boa.fiberby.dk) (Quit: Ex-Chat)</li>
 <li id="l-472"><a href="#l-472">#</a> [18:25] * Joins: smaug____ (~chatzilla@cs181151161.pp.htv.fi)</li>
 <li id="l-473"><a href="#l-473">#</a> [18:31] * Quits: jdong_bot_ (~jdong_bot@118.186.129.129) (Remote host closed the connection)</li>
 <li id="l-474"><a href="#l-474">#</a> [18:38] * Quits: GlitchMr (~glitchmr@178-36-55-53.adsl.inetia.pl) (Quit: GlitchMr)</li>
 <li id="l-475"><a href="#l-475">#</a> [18:38] * Quits: jochen__ (jochen@nat/google/x-khisooetcorhewbl) (Remote host closed the connection)</li>
 <li id="l-476"><a href="#l-476">#</a> [18:39] * Joins: jochen__ (jochen@nat/google/x-segjdhypnhdddbki)</li>
 <li id="l-477" _a="jgraham" _r="dglazkov"><a href="#l-477">#</a> [18:39] &lt;jgraham&gt; dglazkov: Climate, mostly, I think</li>
 <li id="l-478"><a href="#l-478">#</a> [18:39] * Joins: mokush (~quassel@188.24.43.163)</li>
 <li id="l-479"><a href="#l-479">#</a> [18:39] * Joins: astearns (~anonymous@192.150.22.5)</li>
 <li id="l-480"><a href="#l-480">#</a> [18:42] * Joins: dave_levin (dave_levin@nat/google/x-esqpseyxvohzrmup)</li>
 <li id="l-481"><a href="#l-481">#</a> [18:42] * Quits: FlorianX1 (~Florian_S@p4FE2CCFB.dip.t-dialin.net) (Read error: Connection reset by peer)</li>
 <li id="l-482" _a="rabbi1"><a href="#l-482">#</a> [18:55] &lt;rabbi1&gt; how to make a website I view float left ?</li>
 <li id="l-483"><a href="#l-483">#</a> [18:55] * Quits: zdobersek1 (~zan@BSN-142-43-64.dial-up.dsl.siol.net) (Read error: Connection reset by peer)</li>
 <li id="l-484"><a href="#l-484">#</a> [18:55] * Joins: zdobersek (~zan@BSN-142-43-64.dial-up.dsl.siol.net)</li>
 <li id="l-485"><a href="#l-485">#</a> [18:57] * Quits: Areks (~Areks@176.14.214.163) (Ping timeout: 276 seconds)</li>
 <li id="l-486"><a href="#l-486">#</a> [19:00] * Quits: divya (~divyam@219.64.117.145) (Quit: Leaving.)</li>
 <li id="l-487"><a href="#l-487">#</a> [19:04] * Joins: dbaron (~dbaron@nat/mozilla/x-hxyyuocufluuvhuj)</li>
 <li id="l-488"><a href="#l-488">#</a> [19:05] * Quits: rtuin (~rtuin@213.125.175.250) (Quit: Leaving)</li>
 <li id="l-489"><a href="#l-489">#</a> [19:15] * Quits: connrs (~connrs@conners.plus.com) (Quit: Going offline)</li>
 <li id="l-490"><a href="#l-490">#</a> [19:17] * Quits: mpt (mpt@canonical/mpt) (Ping timeout: 256 seconds)</li>
 <li id="l-491"><a href="#l-491">#</a> [19:33] * Joins: connrs (~connrs@conners.plus.com)</li>
 <li id="l-492" _a="AryehGregor" _r="annevk"><a href="#l-492">#</a> [19:37] &lt;AryehGregor&gt; annevk, yes, it didn't mention testharnessreport.js anymore.  jgraham fixed it: <a href="http://dvcs.w3.org/hg/html/rev/bcf81d8686ab">http://dvcs.w3.org/hg/html/rev/bcf81d8686ab</a></li>
 <li id="l-493"><a href="#l-493">#</a> [19:41] * Quits: jwm (~jwm@c-50-129-90-95.hsd1.in.comcast.net) (Remote host closed the connection)</li>
 <li id="l-494" _a="NIFTYLETTUCE"><a href="#l-494">#</a> [19:42] * NIFTYLETTUCE is now known as niftylettuce</li>
 <li id="l-495"><a href="#l-495">#</a> [19:43] * Joins: jwm (~jwm@c-50-129-90-95.hsd1.in.comcast.net)</li>
 <li id="l-496"><a href="#l-496">#</a> [19:46] * Quits: abarth (~abarth@173-164-128-209-SFBA.hfc.comcastbusiness.net) (Quit: abarth)</li>
 <li id="l-497"><a href="#l-497">#</a> [19:52] * Quits: jernoble (~jernoble@17.212.152.13) (Remote host closed the connection)</li>
 <li id="l-498"><a href="#l-498">#</a> [19:56] * Joins: zdobersek3 (~zan@BSN-142-43-64.dial-up.dsl.siol.net)</li>
 <li id="l-499"><a href="#l-499">#</a> [19:56] * Quits: zdobersek (~zan@BSN-142-43-64.dial-up.dsl.siol.net) (Quit: Leaving.)</li>
 <li id="l-500"><a href="#l-500">#</a> [19:58] * Joins: ojan (ojan@nat/google/x-mjuvzrhheztviixx)</li>
 <li id="l-501"><a href="#l-501">#</a> [19:58] * Joins: mpt (mpt@canonical/mpt)</li>
 <li id="l-502"><a href="#l-502">#</a> [19:59] * Joins: hasather_ (~hasather_@84.38.144.96)</li>
 <li id="l-503"><a href="#l-503">#</a> [20:00] * Joins: jernoble (~jernoble@2620:149:4:1b01:8152:be87:6579:5abc)</li>
 <li id="l-504"><a href="#l-504">#</a> [20:00] * Joins: rniwa (rniwa@nat/google/x-wyhyrrfvsarajoyg)</li>
 <li id="l-505" _a="zdobersek3"><a href="#l-505">#</a> [20:02] * zdobersek3 is now known as zdobersek</li>
 <li id="l-506" _a="ericc|afk"><a href="#l-506">#</a> [20:03] * ericc|afk is now known as ericc|mtg</li>
 <li id="l-507"><a href="#l-507">#</a> [20:03] * Quits: jernoble (~jernoble@2620:149:4:1b01:8152:be87:6579:5abc) (Read error: Connection reset by peer)</li>
 <li id="l-508"><a href="#l-508">#</a> [20:04] * Joins: jernoble (~jernoble@17.212.152.13)</li>
 <li id="l-509"><a href="#l-509">#</a> [20:05] * Joins: jwalden (~waldo@2620:101:8003:200:224:d7ff:fef0:8d90)</li>
 <li id="l-510"><a href="#l-510">#</a> [20:08] * Quits: rabbi1 (~manjunath@49.249.126.64) (Quit: Leaving.)</li>
 <li id="l-511"><a href="#l-511">#</a> [20:09] * Joins: GlitchMr (~glitchmr@178-36-55-53.adsl.inetia.pl)</li>
 <li id="l-512"><a href="#l-512">#</a> [20:18] * Quits: rillian_ (~rillian@mist.thaumas.net) (Remote host closed the connection)</li>
 <li id="l-513"><a href="#l-513">#</a> [20:27] * Quits: karlcow (~karl@nerval.la-grange.net) (Quit: Freedom - to walk free and own no superior.)</li>
 <li id="l-514"><a href="#l-514">#</a> [20:29] * Joins: MikeSmith_ (~MikeSmith@EM114-48-48-111.pool.e-mobile.ne.jp)</li>
 <li id="l-515"><a href="#l-515">#</a> [20:29] * Joins: nessy (~Adium@124-168-24-196.dyn.iinet.net.au)</li>
 <li id="l-516"><a href="#l-516">#</a> [20:32] * Quits: annevk (~annevk@EM1-113-75-77.pool.e-mobile.ne.jp) (Ping timeout: 248 seconds)</li>
 <li id="l-517"><a href="#l-517">#</a> [20:32] * Quits: MikeSmith (~MikeSmith@EM1-113-75-77.pool.e-mobile.ne.jp) (Ping timeout: 258 seconds)</li>
 <li id="l-518" _a="MikeSmith_"><a href="#l-518">#</a> [20:32] * MikeSmith_ is now known as MikeSmith</li>
 <li id="l-519"><a href="#l-519">#</a> [20:34] * Quits: mpt (mpt@canonical/mpt) (Ping timeout: 256 seconds)</li>
 <li id="l-520"><a href="#l-520">#</a> [20:35] * Quits: nessy (~Adium@124-168-24-196.dyn.iinet.net.au) (Quit: Leaving.)</li>
 <li id="l-521"><a href="#l-521">#</a> [20:35] * Joins: sicking (~chatzilla@206-15-76-122.static.twtelecom.net)</li>
 <li id="l-522"><a href="#l-522">#</a> [20:42] * Joins: weinig (~weinig@17.212.155.112)</li>
 <li id="l-523" _a="AryehGregor"><a href="#l-523">#</a> [20:43] &lt;AryehGregor&gt; Are init*Event() methods deprecated in new APIs?  If so, what's the correct replacement, event constructors?</li>
 <li id="l-524" _a="Hixie"><a href="#l-524">#</a> [20:43] &lt;Hixie&gt; yes and yes</li>
 <li id="l-525" _a="AryehGregor"><a href="#l-525">#</a> [20:43] &lt;AryehGregor&gt; k.</li>
 <li id="l-526" _a="Hixie"><a href="#l-526">#</a> [20:43] &lt;Hixie&gt; not just deprecated, they shouldn't be added at all</li>
 <li id="l-527" _a="AryehGregor"><a href="#l-527">#</a> [20:43] &lt;AryehGregor&gt; That's what I mean.</li>
 <li id="l-528" _a="AryehGregor"><a href="#l-528">#</a> [20:45] &lt;AryehGregor&gt; Where's an example of event constructors in a spec that I can point to?</li>
 <li id="l-529" _a="Hixie"><a href="#l-529">#</a> [20:45] &lt;Hixie&gt; all of HTML's events are done that way</li>
 <li id="l-530" _a="AryehGregor"><a href="#l-530">#</a> [20:45] &lt;AryehGregor&gt; Okay.</li>
 <li id="l-531" _a="Hixie"><a href="#l-531">#</a> [20:45] &lt;Hixie&gt; search for &quot;: Event {&quot;</li>
 <li id="l-532"><a href="#l-532">#</a> [20:47] * Quits: jwalden (~waldo@2620:101:8003:200:224:d7ff:fef0:8d90) (Ping timeout: 244 seconds)</li>
 <li id="l-533"><a href="#l-533">#</a> [20:50] * Joins: rillian_ (~rillian@mist.thaumas.net)</li>
 <li id="l-534" _a="heycam" _r="Hixie"><a href="#l-534">#</a> [20:51] &lt;heycam&gt; Hixie, you didn't remove the existing init* methods though did you?</li>
 <li id="l-535" _a="Hixie"><a href="#l-535">#</a> [20:51] &lt;Hixie&gt; yes</li>
 <li id="l-536" _a="Hixie"><a href="#l-536">#</a> [20:52] &lt;Hixie&gt; they are gone</li>
 <li id="l-537" _a="heycam"><a href="#l-537">#</a> [20:52] &lt;heycam&gt; even for interfaces that have been deployed for a while?</li>
 <li id="l-538" _a="Hixie"><a href="#l-538">#</a> [20:52] &lt;Hixie&gt; none of the events defined in the html spec are particularly old</li>
 <li id="l-539" _a="Hixie"><a href="#l-539">#</a> [20:52] &lt;Hixie&gt; none of the event interfaces, i should say</li>
 <li id="l-540" _a="heycam" _r="ah"><a href="#l-540">#</a> [20:52] &lt;heycam&gt; ah, ok</li>
 <li id="l-541"><a href="#l-541">#</a> [20:54] * Joins: Areks (~Areks@176.14.214.163)</li>
 <li id="l-542"><a href="#l-542">#</a> [20:57] * Joins: mpt (mpt@canonical/mpt)</li>
 <li id="l-543" _a="AryehGregor" _r="Hixie"><a href="#l-543">#</a> [21:01] &lt;AryehGregor&gt; Hixie, is the &quot;resolve a URL&quot; algorithm in the spec correct/stable?</li>
 <li id="l-544" _a="AryehGregor"><a href="#l-544">#</a> [21:01] &lt;AryehGregor&gt; I seem to recall it wandered around a bit.</li>
 <li id="l-545"><a href="#l-545">#</a> [21:02] * Quits: rillian_ (~rillian@mist.thaumas.net) (Remote host closed the connection)</li>
 <li id="l-546"><a href="#l-546">#</a> [21:05] * Joins: svl (~me@ip565744a7.direct-adsl.nl)</li>
 <li id="l-547" _a="AryehGregor"><a href="#l-547">#</a> [21:07] * AryehGregor determines it's way too complicated to incorporate into reflection tests anyway, bleh</li>
 <li id="l-548"><a href="#l-548">#</a> [21:07] * Quits: mokush (~quassel@188.24.43.163) (Remote host closed the connection)</li>
 <li id="l-549"><a href="#l-549">#</a> [21:14] * Parts: brucel (~brucel@cpc4-smal11-2-0-cust879.perr.cable.virginmedia.com)</li>
 <li id="l-550"><a href="#l-550">#</a> [21:17] * Joins: jwalden (~waldo@2620:101:8003:200:224:d7ff:fef0:8d90)</li>
 <li id="l-551"><a href="#l-551">#</a> [21:17] * Quits: Timz (~Adium@86.89.174.199) (Quit: Leaving.)</li>
 <li id="l-552"><a href="#l-552">#</a> [21:26] * Quits: zdobersek (~zan@BSN-142-43-64.dial-up.dsl.siol.net) (Quit: Leaving.)</li>
 <li id="l-553"><a href="#l-553">#</a> [21:27] * Quits: ojan (ojan@nat/google/x-mjuvzrhheztviixx) (Quit: ojan)</li>
 <li id="l-554"><a href="#l-554">#</a> [21:32] * Quits: tndH (~Rob@cpc16-seac19-2-0-cust234.7-2.cable.virginmedia.com) (Remote host closed the connection)</li>
 <li id="l-555"><a href="#l-555">#</a> [21:34] * Joins: nielsle_ (~nielsle@3239059-cl69.boa.fiberby.dk)</li>
 <li id="l-556"><a href="#l-556">#</a> [21:40] * Joins: bga_ (~bga@ppp78-37-214-230.pppoe.avangarddsl.ru)</li>
 <li id="l-557"><a href="#l-557">#</a> [21:40] * Quits: _bga (~bga@ppp78-37-214-230.pppoe.avangarddsl.ru) (Read error: Connection reset by peer)</li>
 <li id="l-558"><a href="#l-558">#</a> [21:46] * Joins: rillian_ (~rillian@mist.thaumas.net)</li>
 <li id="l-559"><a href="#l-559">#</a> [21:54] * Quits: jwalden (~waldo@2620:101:8003:200:224:d7ff:fef0:8d90) (Ping timeout: 240 seconds)</li>
 <li id="l-560" _a="TabAtkins" _r="Hm"><a href="#l-560">#</a> [21:57] &lt;TabAtkins&gt; Hm, it appears that all of my windows have lost their decorations.  Time to reboot, I guess.</li>
 <li id="l-561"><a href="#l-561">#</a> [21:57] * Quits: TabAtkins (tabatkins@nat/google/x-nsvkozmxwfrmlkjv) (Quit: Lost terminal)</li>
 <li id="l-562" _a="bga_"><a href="#l-562">#</a> [21:58] &lt;bga_&gt; <a href="http://mobile.twitter.com/wendyg/status/128554733714669568">http://mobile.twitter.com/wendyg/status/128554733714669568</a></li>
 <li id="l-563" _a="paul_irish"><a href="#l-563">#</a> [21:58] &lt;paul_irish&gt; &quot;R.I.P. John McCarthy, father of AI, inventor of Lisp, suddenly at home last night. Pls RT.&quot;</li>
 <li id="l-564"><a href="#l-564">#</a> [22:02] * Quits: weinig (~weinig@17.212.155.112) (Quit: weinig)</li>
 <li id="l-565"><a href="#l-565">#</a> [22:02] * Joins: jwalden (~waldo@2620:101:8003:200:224:d7ff:fef0:8d90)</li>
 <li id="l-566"><a href="#l-566">#</a> [22:03] * Joins: nessy (~Adium@124-168-24-196.dyn.iinet.net.au)</li>
 <li id="l-567"><a href="#l-567">#</a> [22:04] * Quits: cpearce (~chatzilla@ip-118-90-78-13.xdsl.xnet.co.nz) (Ping timeout: 240 seconds)</li>
 <li id="l-568" _a="AryehGregor" _r="jgraham"><a href="#l-568">#</a> [22:04] &lt;AryehGregor&gt; jgraham, what should we do when doing assert_throws() with new exception types, like INVALID_NODE_TYPE_ERR?</li>
 <li id="l-569"><a href="#l-569">#</a> [22:04] * Joins: jamesr (jamesr@nat/google/x-blndjvsxvjzzcglr)</li>
 <li id="l-570" _a="AryehGregor" _r="Oh"><a href="#l-570">#</a> [22:05] &lt;AryehGregor&gt; Oh, wait, never mind.</li>
 <li id="l-571" _a="AryehGregor"><a href="#l-571">#</a> [22:05] &lt;AryehGregor&gt; It was a typo on my part.</li>
 <li id="l-572" _a="AryehGregor"><a href="#l-572">#</a> [22:06] &lt;AryehGregor&gt; I mean, the current testharness.js still doesn't match the spec, but that can be fixed.</li>
 <li id="l-573"><a href="#l-573">#</a> [22:06] * Joins: weinig (~weinig@17.245.88.60)</li>
 <li id="l-574"><a href="#l-574">#</a> [22:07] * Quits: weinig (~weinig@17.245.88.60) (Client Quit)</li>
 <li id="l-575"><a href="#l-575">#</a> [22:08] * Quits: dbaron (~dbaron@nat/mozilla/x-hxyyuocufluuvhuj) (Quit: 8403864 bytes have been tenured, next gc will be global.)</li>
 <li id="l-576"><a href="#l-576">#</a> [22:10] * Joins: dbaron (~dbaron@nat/mozilla/x-bokcvjiyjptqjbdd)</li>
 <li id="l-577"><a href="#l-577">#</a> [22:11] * Quits: GlitchMr (~glitchmr@178-36-55-53.adsl.inetia.pl) (Read error: Connection reset by peer)</li>
 <li id="l-578"><a href="#l-578">#</a> [22:14] * Quits: nessy (~Adium@124-168-24-196.dyn.iinet.net.au) (Quit: Leaving.)</li>
 <li id="l-579"><a href="#l-579">#</a> [22:14] * Quits: KillerX (~anant@nat/mozilla/x-yftvkpxschruugkb) (Ping timeout: 260 seconds)</li>
 <li id="l-580"><a href="#l-580">#</a> [22:15] * Quits: crash\ (bouncer@lubyte.de) (Ping timeout: 248 seconds)</li>
 <li id="l-581"><a href="#l-581">#</a> [22:15] * Quits: jwalden (~waldo@2620:101:8003:200:224:d7ff:fef0:8d90) (Ping timeout: 244 seconds)</li>
 <li id="l-582"><a href="#l-582">#</a> [22:16] * Quits: jamesr (jamesr@nat/google/x-blndjvsxvjzzcglr) (Ping timeout: 244 seconds)</li>
 <li id="l-583"><a href="#l-583">#</a> [22:16] * Quits: dbaron (~dbaron@nat/mozilla/x-bokcvjiyjptqjbdd) (Ping timeout: 255 seconds)</li>
 <li id="l-584"><a href="#l-584">#</a> [22:16] * Joins: TabAtkins (tabatkins@nat/google/x-tikwitgbqveakztl)</li>
 <li id="l-585"><a href="#l-585">#</a> [22:17] * Joins: roc (~chatzilla@60.234.54.74)</li>
 <li id="l-586"><a href="#l-586">#</a> [22:19] * Quits: scor (~scor@drupal.org/user/52142/view) (Quit: scor)</li>
 <li id="l-587"><a href="#l-587">#</a> [22:20] * Joins: weinig (~weinig@17.245.88.60)</li>
 <li id="l-588"><a href="#l-588">#</a> [22:29] * Joins: ojan (ojan@nat/google/x-djvdpcpeoefopzzp)</li>
 <li id="l-589"><a href="#l-589">#</a> [22:29] * Quits: smaug____ (~chatzilla@cs181151161.pp.htv.fi) (Ping timeout: 258 seconds)</li>
 <li id="l-590"><a href="#l-590">#</a> [22:32] * Joins: nessy (~Adium@124-168-24-196.dyn.iinet.net.au)</li>
 <li id="l-591"><a href="#l-591">#</a> [22:32] * Joins: jwalden (~waldo@2620:101:8003:200:224:d7ff:fef0:8d90)</li>
 <li id="l-592"><a href="#l-592">#</a> [22:36] * Quits: miketaylr (~miketaylr@206.217.92.186) (Quit: miketaylr)</li>
 <li id="l-593"><a href="#l-593">#</a> [22:39] * Quits: weinig (~weinig@17.245.88.60) (Quit: weinig)</li>
 <li id="l-594"><a href="#l-594">#</a> [22:41] * Joins: crash\ (bouncer@lubyte.de)</li>
 <li id="l-595"><a href="#l-595">#</a> [22:41] * Quits: kennyluck (~kennyluck@114-43-119-208.dynamic.hinet.net) (Read error: Connection reset by peer)</li>
 <li id="l-596"><a href="#l-596">#</a> [22:44] * Joins: weinig (~weinig@17.245.88.60)</li>
 <li id="l-597"><a href="#l-597">#</a> [22:45] * Quits: davidb_ (~davidb@66.207.208.98) (Quit: davidb_)</li>
 <li id="l-598"><a href="#l-598">#</a> [22:48] * Quits: crash\ (bouncer@lubyte.de) (Ping timeout: 240 seconds)</li>
 <li id="l-599"><a href="#l-599">#</a> [22:48] * Joins: kennyluck (~kennyluck@114-43-124-16.dynamic.hinet.net)</li>
 <li id="l-600"><a href="#l-600">#</a> [22:49] * Joins: crash\ (bouncer@lubyte.de)</li>
 <li id="l-601"><a href="#l-601">#</a> [22:49] * Quits: tomasf (~tom@c-5ed9e555.024-204-6c6b7012.cust.bredbandsbolaget.se) (Quit: tomasf)</li>
 <li id="l-602"><a href="#l-602">#</a> [22:50] * Quits: weinig (~weinig@17.245.88.60) (Quit: weinig)</li>
 <li id="l-603" _a="ericc|mtg"><a href="#l-603">#</a> [22:57] * ericc|mtg is now known as eric_carlson</li>
 <li id="l-604"><a href="#l-604">#</a> [23:07] * Quits: jwalden (~waldo@2620:101:8003:200:224:d7ff:fef0:8d90) (Ping timeout: 244 seconds)</li>
 <li id="l-605"><a href="#l-605">#</a> [23:08] * Joins: KillerX (~anant@nat/mozilla/x-gxfmcbqwipblzrvu)</li>
 <li id="l-606" _a="rniwa" _r="AryehGregor"><a href="#l-606">#</a> [23:08] &lt;rniwa&gt; AryehGregor: yt?</li>
 <li id="l-607"><a href="#l-607">#</a> [23:10] * Quits: othermaciej_ (~mjs@c-24-6-209-189.hsd1.ca.comcast.net) (Quit: othermaciej_)</li>
 <li id="l-608" _a="zewt"><a href="#l-608">#</a> [23:11] &lt;zewt&gt; chrome accepts xhr.responseType = &quot;blob&quot; but xhr.response comes back as null? :(</li>
 <li id="l-609" _a="zewt"><a href="#l-609">#</a> [23:11] &lt;zewt&gt; (maybe I'm doing something wrong, but if they don't support it yet, it shouldn't accept &quot;blob&quot; in the first place)</li>
 <li id="l-610"><a href="#l-610">#</a> [23:11] * Joins: smaug____ (~chatzilla@cs181151161.pp.htv.fi)</li>
 <li id="l-611"><a href="#l-611">#</a> [23:11] * Quits: MacTed (~Thud@63.119.36.36)</li>
 <li id="l-612" _a="AryehGregor" _r="rniwa"><a href="#l-612">#</a> [23:12] &lt;AryehGregor&gt; rniwa, yep.</li>
 <li id="l-613" _a="AryehGregor"><a href="#l-613">#</a> [23:12] &lt;AryehGregor&gt; (not for long, though)</li>
 <li id="l-614"><a href="#l-614">#</a> [23:12] * Joins: jwalden (~waldo@2620:101:8003:200:224:d7ff:fef0:8d90)</li>
 <li id="l-615" _a="AryehGregor"><a href="#l-615">#</a> [23:13] &lt;AryehGregor&gt; (around half an hour)</li>
 <li id="l-616" _a="rniwa" _r="AryehGregor"><a href="#l-616">#</a> [23:14] &lt;rniwa&gt; AryehGregor: k</li>
 <li id="l-617"><a href="#l-617">#</a> [23:14] * Joins: dbaron (~dbaron@nat/mozilla/x-cxlxpiculprusrnt)</li>
 <li id="l-618" _a="rniwa" _r="AryehGregor"><a href="#l-618">#</a> [23:14] &lt;rniwa&gt; AryehGregor: thanks for responding to my emails :)</li>
 <li id="l-619" _a="AryehGregor" _r="rniwa"><a href="#l-619">#</a> [23:14] &lt;AryehGregor&gt; rniwa, I'll write some more tomorrow.</li>
 <li id="l-620" _a="rniwa" _r="AryehGregor"><a href="#l-620">#</a> [23:14] &lt;rniwa&gt; AryehGregor: I totally understand the delays and all that jazz so take your time :)</li>
 <li id="l-621" _a="rniwa" _r="AryehGregor"><a href="#l-621">#</a> [23:14] &lt;rniwa&gt; AryehGregor: especially given that you're no longer compensated for your work</li>
 <li id="l-622" _a="AryehGregor" _r="rniwa"><a href="#l-622">#</a> [23:14] &lt;AryehGregor&gt; rniwa, actually, I am.</li>
 <li id="l-623" _a="rniwa"><a href="#l-623">#</a> [23:14] &lt;rniwa&gt; oh good!</li>
 <li id="l-624" _a="AryehGregor"><a href="#l-624">#</a> [23:14] &lt;AryehGregor&gt; If I weren't, I wouldn't be doing it.  :)</li>
 <li id="l-625" _a="rniwa"><a href="#l-625">#</a> [23:14] &lt;rniwa&gt; glad to hear that</li>
 <li id="l-626" _a="AryehGregor"><a href="#l-626">#</a> [23:14] &lt;AryehGregor&gt; (but I expect to continue to be compensated by someone)</li>
 <li id="l-627" _a="rniwa" _r="okay"><a href="#l-627">#</a> [23:15] &lt;rniwa&gt; okay, I feel less guilty to delegate more work to you then :P</li>
 <li id="l-628" _a="rniwa" _r="AryehGregor"><a href="#l-628">#</a> [23:15] &lt;rniwa&gt; AryehGregor: anyway, the immediate concern for me is whether my spec should define what undo/redo events do or not</li>
 <li id="l-629" _a="AryehGregor" _r="rniwa"><a href="#l-629">#</a> [23:16] &lt;AryehGregor&gt; rniwa, it would make sense.</li>
 <li id="l-630" _a="rniwa"><a href="#l-630">#</a> [23:16] &lt;rniwa&gt; ok</li>
 <li id="l-631" _a="rniwa" _r="AryehGregor"><a href="#l-631">#</a> [23:16] &lt;rniwa&gt; AryehGregor: the biggest problem I have with how selection is restored on undo/redo</li>
 <li id="l-632" _a="rniwa" _r="AryehGregor"><a href="#l-632">#</a> [23:17] &lt;rniwa&gt; AryehGregor: also, there are many selection behaviors that can't be cross-plaform</li>
 <li id="l-633" _a="rniwa"><a href="#l-633">#</a> [23:17] &lt;rniwa&gt; e.g. extending between word boundaries on Windows and Mac use different conventions to include/exclude trailing whitespace</li>
 <li id="l-634" _a="rniwa"><a href="#l-634">#</a> [23:17] &lt;rniwa&gt; so those things probably need to be investigated</li>
 <li id="l-635" _a="AryehGregor"><a href="#l-635">#</a> [23:18] &lt;AryehGregor&gt; Yes.</li>
 <li id="l-636" _a="rniwa"><a href="#l-636">#</a> [23:18] &lt;rniwa&gt; though I don't think it'll affect undo/redo that much</li>
 <li id="l-637" _a="rniwa"><a href="#l-637">#</a> [23:18] &lt;rniwa&gt; except the one case I pointed out which is whether or not the deleted content should be selected after undo</li>
 <li id="l-638"><a href="#l-638">#</a> [23:20] * Joins: micheil (~micheil@92.40.254.106.threembb.co.uk)</li>
 <li id="l-639"><a href="#l-639">#</a> [23:20] * Quits: benjoffe_ (~benjoffe_@CPE-121-218-141-115.lnse3.cht.bigpond.net.au) (Remote host closed the connection)</li>
 <li id="l-640"><a href="#l-640">#</a> [23:22] * Quits: Maurice (copyman@5ED573FA.cm-7-6b.dynamic.ziggo.nl)</li>
 <li id="l-641"><a href="#l-641">#</a> [23:30] * Quits: nielsle_ (~nielsle@3239059-cl69.boa.fiberby.dk) (Ping timeout: 276 seconds)</li>
 <li id="l-642"><a href="#l-642">#</a> [23:39] * Joins: slightlyoff_afk (u1768@gateway/web/irccloud.com/x-zgkucfuqtkxyikfq)</li>
 <li id="l-643" _a="slightlyoff_afk"><a href="#l-643">#</a> [23:39] * slightlyoff_afk is now known as slightlyoff</li>
 <li id="l-644" _a="eric_carlson"><a href="#l-644">#</a> [23:52] * eric_carlson is now known as ericc|afk</li>
 <li id="l-645"><a href="#l-645">#</a> [23:54] * Quits: svl (~me@ip565744a7.direct-adsl.nl) (Quit: And back he spurred like a madman, shrieking a curse to the sky.)</li>
 <li id="l-646"><a href="#l-646">#</a> [23:58] * Joins: erlehmann (~erlehmann@82.113.121.160)</li>
 <li id="l-647"><a href="#l-647">#</a> [23:59] * Quits: smaug____ (~chatzilla@cs181151161.pp.htv.fi) (Ping timeout: 258 seconds)</li>
 <li id="l-648"><a href="#l-648">#</a> Session Close: Tue Oct 25 00:00:00 2011</li>
</ol>
<p id="end">The end :)
